refactor: refactor app state to make it less of a monster
