+++ lib/state/app_state.dart
@@ -46,6 +46,12 @@ import 'sidebar_item.dart';
 import 'theme_palette_source.dart';
 import 'track_list_style.dart';
 
+part 'app_state_session.dart';
+part 'app_state_library.dart';
+part 'app_state_preferences.dart';
+part 'app_state_offline.dart';
+part 'app_state_favorites.dart';
+
 const Set<ConnectivityResult> _networkConnectivityWhitelist = {
   ConnectivityResult.wifi,
   ConnectivityResult.ethernet,
@@ -95,6 +101,10 @@ class AppState extends ChangeNotifier {
 
   void _notifyListenersLater() {
     SchedulerBinding.instance.addPostFrameCallback((_) => notifyListeners());
+  }
+
+  void _notify() {
+    notifyListeners();
   }
 
   List<Playlist> _playlists = [];
@@ -830,718 +840,12 @@ class AppState extends ChangeNotifier {
   bool isSidebarItemVisible(SidebarItem item) =>
       _sidebarVisibility[item] ?? true;
 
-  /// Updates the browse layout for a library view.
-  void setBrowseLayout(LibraryView view, BrowseLayout layout) {
-    _browseLayouts[view] = layout;
-    notifyListeners();
-  }
-
-  /// Updates the visibility of a home section.
-  Future<void> setHomeSectionVisible(
-    HomeSection section,
-    bool visible,
-  ) async {
-    _homeSectionVisibility[section] = visible;
-    await _settingsStore.saveHomeSectionVisibility(_homeSectionVisibility);
-    notifyListeners();
-    if (section == HomeSection.jumpIn && visible) {
-      unawaited(loadJumpIn(force: true));
-    }
-  }
-
-  /// Updates the ordering of home sections.
-  Future<void> reorderHomeSections(int oldIndex, int newIndex) async {
-    if (oldIndex == newIndex) {
-      return;
-    }
-    final list = List<HomeSection>.from(_homeSectionOrder);
-    if (oldIndex < 0 || oldIndex >= list.length) {
-      return;
-    }
-    if (newIndex > oldIndex) {
-      newIndex -= 1;
-    }
-    final moved = list.removeAt(oldIndex);
-    final target = newIndex.clamp(0, list.length);
-    list.insert(target, moved);
-    _homeSectionOrder = list;
-    await _settingsStore.saveHomeSectionOrder(_homeSectionOrder);
-    notifyListeners();
-  }
-
-  /// Updates the visibility of a sidebar item.
-  Future<void> setSidebarItemVisible(
-    SidebarItem item,
-    bool visible,
-  ) async {
-    _sidebarVisibility[item] = visible;
-    await _settingsStore.saveSidebarVisibility(_sidebarVisibility);
-    notifyListeners();
-  }
-
   /// Returns a saved scroll offset for a key.
   double loadScrollOffset(String key) => _scrollOffsets[key] ?? 0;
 
   /// Saves a scroll offset for a key.
   void saveScrollOffset(String key, double offset) {
     _scrollOffsets[key] = offset;
-  }
-
-  /// Initializes cached state and refreshes library.
-  Future<void> bootstrap() async {
-    await AppInfo.load();
-    final deviceId = await _settingsStore.loadDeviceId();
-    _client.updateDeviceInfo(
-      deviceId: deviceId,
-      deviceName: _platformDeviceName(),
-    );
-    _session = await _sessionStore.loadSession();
-    if (_session != null) {
-      _client.updateSession(_session!);
-    }
-    _themeMode = await _settingsStore.loadThemeMode();
-    _fontFamily = await _settingsStore.loadFontFamily();
-    _fontScale = await _settingsStore.loadFontScale();
-    _accentColorValue = await _settingsStore.loadAccentColorValue();
-    _accentColorSource = await _settingsStore.loadAccentColorSource();
-    _themePaletteSource = await _settingsStore.loadThemePaletteSource();
-    _telemetryPlayback = await _settingsStore.loadPlaybackTelemetry();
-    _telemetryProgress = await _settingsStore.loadProgressTelemetry();
-    _telemetryHistory = await _settingsStore.loadHistoryTelemetry();
-    _gaplessPlayback = await _settingsStore.loadGaplessPlayback();
-    _downloadsWifiOnly = await _settingsStore.loadDownloadsWifiOnly();
-    _downloadsPaused = await _settingsStore.loadDownloadsPaused();
-    _autoDownloadFavoritesEnabled =
-        await _settingsStore.loadAutoDownloadFavoritesEnabled();
-    _autoDownloadFavoriteAlbums =
-        await _settingsStore.loadAutoDownloadFavoriteAlbums();
-    _autoDownloadFavoriteArtists =
-        await _settingsStore.loadAutoDownloadFavoriteArtists();
-    _autoDownloadFavoriteTracks =
-        await _settingsStore.loadAutoDownloadFavoriteTracks();
-    _autoDownloadFavoritesWifiOnly =
-        await _settingsStore.loadAutoDownloadFavoritesWifiOnly();
-    _settingsShortcutEnabled =
-        await _settingsStore.loadSettingsShortcutEnabled();
-    _settingsShortcut = await _settingsStore.loadSettingsShortcut();
-    _searchShortcutEnabled = await _settingsStore.loadSearchShortcutEnabled();
-    _searchShortcut = await _settingsStore.loadSearchShortcut();
-    _preferLocalSearch = await _settingsStore.loadPreferLocalSearch();
-    _layoutDensity = await _settingsStore.loadLayoutDensity();
-    _cornerRadiusStyle = await _settingsStore.loadCornerRadiusStyle();
-    _trackListStyle = await _settingsStore.loadTrackListStyle();
-    _trackStatusIconsEnabled =
-        await _settingsStore.loadTrackStatusIconsEnabled();
-    _nowPlayingLayout = await _settingsStore.loadNowPlayingLayout();
-    _homeShelfLayout = await _settingsStore.loadHomeShelfLayout();
-    _homeShelfGridRows = await _settingsStore.loadHomeShelfGridRows();
-    _offlineMode = await _settingsStore.loadOfflineMode();
-    _cacheMaxBytes = await _cacheStore.loadCacheMaxBytes();
-    _homeSectionVisibility = await _settingsStore.loadHomeSectionVisibility();
-    _homeSectionOrder = await _settingsStore.loadHomeSectionOrder();
-    _sidebarVisibility = await _settingsStore.loadSidebarVisibility();
-    _sidebarWidth = await _settingsStore.loadSidebarWidth();
-    _sidebarCollapsed = await _settingsStore.loadSidebarCollapsed();
-    _smartLists = await _settingsStore.loadSmartLists();
-    _pinnedAudio = await _cacheStore.loadPinnedAudio();
-    _ensureHomeInHistory();
-    unawaited(refreshMediaCacheBytes());
-    await _loadCachedLibrary();
-    await _applyPlaybackSettings();
-    if (_offlineMode) {
-      await _applyOfflineModeData();
-    }
-    await _restorePlaybackResumeState();
-    unawaited(_maybeUpdateNowPlayingPalette(_nowPlaying));
-    _isBootstrapping = false;
-    notifyListeners();
-
-    if (_session != null && !_offlineMode) {
-      await refreshLibrary();
-    }
-  }
-
-  Future<void> _applyPlaybackSettings() async {
-    await _playback.setGaplessPlayback(_gaplessPlayback);
-  }
-
-  /// Attempts Jellyfin sign-in.
-  Future<bool> signIn({
-    required String serverUrl,
-    required String username,
-    required String password,
-  }) async {
-    _authError = null;
-    notifyListeners();
-    try {
-      final session = await _client.authenticate(
-        serverUrl: serverUrl,
-        username: username,
-        password: password,
-      );
-      _session = session;
-      await _sessionStore.saveSession(session);
-      await refreshLibrary();
-      notifyListeners();
-      return true;
-    } catch (error, stackTrace) {
-      final logService = await LogService.instance;
-      await logService.error('Sign in failed', error, stackTrace);
-      _authError = error.toString();
-      notifyListeners();
-      return false;
-    }
-  }
-
-  /// Signs out and clears cached state.
-  Future<void> signOut() async {
-    await _cacheStore.savePlaybackResumeState(null);
-    _session = null;
-    _client.clearSession();
-    _selectedPlaylist = null;
-    _selectedSmartList = null;
-    _selectedView = LibraryView.home;
-    _viewHistory.clear();
-    _ensureHomeInHistory();
-    _selectedAlbum = null;
-    _selectedArtist = null;
-    _selectedGenre = null;
-    _searchQuery = '';
-    _searchResults = null;
-    _isSearching = false;
-    _isSearchLoading = false;
-    _playlistTracks = [];
-    _smartListTracks = [];
-    _featuredTracks = [];
-    _playlists = [];
-    _albums = [];
-    _artists = [];
-    _genres = [];
-    _libraryTracks = [];
-    _trackBrowseLetter = null;
-    _tracksLoadCompleter = null;
-    _isLoadingTracks = false;
-    _hasMoreTracks = true;
-    _tracksOffset = 0;
-    _libraryTracksFromOfflineSnapshot = false;
-    _albumTracks = [];
-    _artistTracks = [];
-    _genreTracks = [];
-    _favoriteAlbums = [];
-    _favoriteArtists = [];
-    _favoriteTracks = [];
-    _recentTracks = [];
-    _playHistory = [];
-    _libraryStats = null;
-    _jumpInTrack = null;
-    _jumpInAlbum = null;
-    _jumpInArtist = null;
-    _isLoadingJumpIn = false;
-    _lastJumpInRefreshAt = null;
-    _queue = [];
-    _downloadQueue.clear();
-    _downloadStatusByUrl.clear();
-    _cancelledOfflineRequests.clear();
-    _isProcessingDownloads = false;
-    _resetPlaybackRuntimeState(clearNowPlaying: true, clearReporting: true);
-    _lastPlaybackPersistAt = null;
-    unawaited(_nowPlayingService.clear());
-    await _sessionStore.saveSession(null);
-    notifyListeners();
-  }
-
-  /// Refreshes playlists and featured tracks.
-  Future<void> refreshLibrary() async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      await _applyOfflineModeData();
-      return;
-    }
-    _isLoadingLibrary = true;
-    notifyListeners();
-    try {
-      final playlists = await _client.fetchPlaylists();
-      _playlists = playlists;
-      await _cacheStore.savePlaylists(playlists);
-      final stats = await _client.fetchLibraryStats();
-      _libraryStats = stats;
-      await _cacheStore.saveLibraryStats(stats);
-      List<MediaItem> recent;
-      try {
-        recent = await _client.fetchRecentlyPlayedTracks();
-      } catch (_) {
-        recent = await _client.fetchRecentTracks();
-      }
-      _recentTracks = recent;
-      await _cacheStore.saveRecentTracks(recent);
-      final featured = await _client.fetchRecentTracks();
-      _featuredTracks = featured;
-      await _cacheStore.saveFeaturedTracks(featured);
-
-      // Load albums and artists eagerly for local search
-      await _loadAlbums();
-      await _loadArtists();
-      await _loadGenres();
-
-      await _loadFavoriteAlbums();
-      await _loadFavoriteArtists();
-      await _loadFavoriteTracks();
-      if (_selectedSmartList != null) {
-        unawaited(_loadSmartListTracks(_selectedSmartList!));
-      }
-      if (isHomeSectionVisible(HomeSection.jumpIn)) {
-        unawaited(loadJumpIn(force: true));
-      }
-    } catch (_) {
-      // Keep cached content if refresh fails.
-    }
-    _isLoadingLibrary = false;
-    notifyListeners();
-  }
-
-  /// Selects a playlist and loads its tracks.
-  Future<void> selectPlaylist(
-    Playlist playlist, {
-    bool offlineOnly = false,
-  }) async {
-    if (_selectedView != LibraryView.home) {
-      _recordViewHistory(_selectedView);
-    }
-    _selectedPlaylist = playlist;
-    _selectedSmartList = null;
-    _selectedView = LibraryView.home;
-    _offlineOnlyFilter = offlineOnly;
-    clearBrowseSelection(notify: false);
-    clearSearch(notify: false);
-    notifyListeners();
-    final cached = await _cacheStore.loadPlaylistTracks(playlist.id);
-    if (cached.isNotEmpty) {
-      _playlistTracks = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      final tracks = await _client.fetchPlaylistTracks(playlist.id);
-      _playlistTracks = tracks;
-      await _cacheStore.savePlaylistTracks(playlist.id, tracks);
-      notifyListeners();
-    } catch (_) {
-      // Keep cached tracks if refresh fails.
-    }
-  }
-
-  /// Loads a playlist and starts playback without navigating.
-  Future<void> playPlaylist(Playlist playlist) async {
-    final logService = await LogService.instance;
-    await logService.info(
-        'playPlaylist: Starting "${playlist.name}" (${playlist.id}), offline=$_offlineMode');
-
-    List<MediaItem> tracks = const [];
-    if (_offlineMode) {
-      await logService
-          .info('playPlaylist: Loading cached tracks for offline mode');
-      tracks = await _cacheStore.loadPlaylistTracks(playlist.id);
-      final filtered = _filterPinnedTracks(tracks);
-      if (filtered.isEmpty) {
-        await logService.warning(
-            'playPlaylist: No pinned tracks available in offline mode');
-        return;
-      }
-      await logService
-          .info('playPlaylist: Playing ${filtered.length} pinned tracks');
-      await _playFromList(filtered, filtered.first);
-      return;
-    }
-    try {
-      await logService.info('playPlaylist: Fetching tracks from server');
-      tracks = await _client.fetchPlaylistTracks(playlist.id);
-      await logService
-          .info('playPlaylist: Fetched ${tracks.length} tracks, caching');
-      await _cacheStore.savePlaylistTracks(playlist.id, tracks);
-    } catch (error, stackTrace) {
-      await logService.error(
-          'playPlaylist: Failed to fetch from server, trying cache',
-          error,
-          stackTrace);
-      tracks = await _cacheStore.loadPlaylistTracks(playlist.id);
-    }
-    if (tracks.isEmpty) {
-      await logService.warning('playPlaylist: No tracks available');
-      return;
-    }
-    await _playFromList(tracks, tracks.first);
-  }
-
-  /// Builds and plays a Smart List without navigating.
-  Future<void> playSmartList(SmartList list) async {
-    await _ensureSmartListSourceLoaded();
-    final tracks = _buildSmartListTracks(list);
-    if (tracks.isEmpty) {
-      return;
-    }
-    await _playFromList(tracks, tracks.first);
-  }
-
-  /// Clears the current playlist selection.
-  void clearPlaylistSelection() {
-    _selectedPlaylist = null;
-    _playlistTracks = [];
-    _selectedView = LibraryView.home;
-    clearBrowseSelection(notify: false);
-    notifyListeners();
-  }
-
-  /// Selects a Smart List and loads its tracks.
-  Future<void> selectSmartList(SmartList list) async {
-    if (_selectedView != LibraryView.home) {
-      _recordViewHistory(_selectedView);
-    }
-    _selectedSmartList = list;
-    _selectedView = LibraryView.home;
-    _offlineOnlyFilter = false;
-    _selectedPlaylist = null;
-    _playlistTracks = [];
-    clearBrowseSelection(notify: false);
-    clearSearch(notify: false);
-    notifyListeners();
-    await _loadSmartListTracks(list);
-  }
-
-  /// Clears the current Smart List selection.
-  void clearSmartListSelection() {
-    _selectedSmartList = null;
-    _smartListTracks = [];
-    _selectedView = LibraryView.home;
-    clearBrowseSelection(notify: false);
-    notifyListeners();
-  }
-
-  /// Creates and stores a Smart List.
-  Future<SmartList> createSmartList(SmartList list) async {
-    _smartLists = [..._smartLists, list]
-      ..sort((a, b) => a.name.toLowerCase().compareTo(b.name.toLowerCase()));
-    await _settingsStore.saveSmartLists(_smartLists);
-    notifyListeners();
-    return list;
-  }
-
-  /// Updates a Smart List definition.
-  Future<void> updateSmartList(SmartList list) async {
-    final index = _smartLists.indexWhere((entry) => entry.id == list.id);
-    if (index == -1) {
-      return;
-    }
-    _smartLists = List<SmartList>.from(_smartLists);
-    _smartLists[index] = list;
-    _smartLists.sort(
-      (a, b) => a.name.toLowerCase().compareTo(b.name.toLowerCase()),
-    );
-    await _settingsStore.saveSmartLists(_smartLists);
-    if (_selectedSmartList?.id == list.id) {
-      _selectedSmartList = list;
-      await _loadSmartListTracks(list);
-    }
-    notifyListeners();
-  }
-
-  /// Deletes a Smart List.
-  Future<void> deleteSmartList(SmartList list) async {
-    _smartLists = _smartLists.where((entry) => entry.id != list.id).toList();
-    await _settingsStore.saveSmartLists(_smartLists);
-    if (_selectedSmartList?.id == list.id) {
-      clearSmartListSelection();
-    } else {
-      notifyListeners();
-    }
-  }
-
-  /// Creates a new playlist.
-  Future<Playlist?> createPlaylist({
-    required String name,
-    List<MediaItem> initialTracks = const [],
-  }) async {
-    if (_session == null || _offlineMode) {
-      return null;
-    }
-    final trimmed = name.trim();
-    if (trimmed.isEmpty) {
-      return null;
-    }
-    try {
-      final playlist = await _client.createPlaylist(
-        name: trimmed,
-        itemIds: initialTracks.map((track) => track.id).toList(),
-      );
-      final created = playlist.id.isEmpty
-          ? Playlist(
-              id: playlist.id,
-              name: trimmed,
-              trackCount: initialTracks.length,
-              imageUrl: playlist.imageUrl,
-            )
-          : playlist;
-      _playlists = [..._playlists, created]..sort(_comparePlaylists);
-      await _cacheStore.savePlaylists(_playlists);
-      _updatePlaylistStats(1);
-      _notifyListenersLater();
-      if (created.id.isNotEmpty && initialTracks.isNotEmpty) {
-        final tracks = await _client.fetchPlaylistTracks(created.id);
-        await _cacheStore.savePlaylistTracks(created.id, tracks);
-        if (_selectedPlaylist?.id == created.id) {
-          _playlistTracks = tracks;
-          notifyListeners();
-        }
-      }
-      return created;
-    } catch (_) {
-      return null;
-    }
-  }
-
-  /// Renames an existing playlist.
-  Future<String?> renamePlaylist(Playlist playlist, String name) async {
-    if (_session == null || _offlineMode) {
-      return 'Playlists are unavailable offline.';
-    }
-    final trimmed = name.trim();
-    if (trimmed.isEmpty || trimmed == playlist.name) {
-      return null;
-    }
-    final updated = Playlist(
-      id: playlist.id,
-      name: trimmed,
-      trackCount: playlist.trackCount,
-      imageUrl: playlist.imageUrl,
-    );
-    final previous = _playlists;
-    _playlists = _playlists
-        .map((item) => item.id == playlist.id ? updated : item)
-        .toList()
-      ..sort(_comparePlaylists);
-    await _cacheStore.savePlaylists(_playlists);
-    notifyListeners();
-    try {
-      await _client.renamePlaylist(playlistId: playlist.id, name: trimmed);
-      if (_selectedPlaylist?.id == playlist.id) {
-        _selectedPlaylist = updated;
-        notifyListeners();
-      }
-      return null;
-    } catch (error) {
-      _playlists = previous;
-      await _cacheStore.savePlaylists(_playlists);
-      notifyListeners();
-      return _requestErrorMessage(
-        error,
-        fallback: 'Unable to rename playlist.',
-      );
-    }
-  }
-
-  /// Deletes a playlist.
-  Future<String?> deletePlaylist(Playlist playlist) async {
-    if (_session == null || _offlineMode) {
-      return 'Playlists are unavailable offline.';
-    }
-    final previous = _playlists;
-    _playlists = _playlists.where((item) => item.id != playlist.id).toList();
-    await _cacheStore.savePlaylists(_playlists);
-    _updatePlaylistStats(-1);
-    if (_selectedPlaylist?.id == playlist.id) {
-      clearPlaylistSelection();
-    } else {
-      notifyListeners();
-    }
-    try {
-      await _client.deletePlaylist(playlist.id);
-      return null;
-    } catch (error) {
-      _playlists = previous;
-      await _cacheStore.savePlaylists(_playlists);
-      _updatePlaylistStats(1);
-      notifyListeners();
-      return _requestErrorMessage(
-        error,
-        fallback: 'Unable to delete playlist.',
-      );
-    }
-  }
-
-  /// Adds a track to the selected playlist.
-  Future<String?> addTrackToPlaylist(
-    MediaItem track,
-    Playlist playlist,
-  ) async {
-    return addTracksToPlaylist(playlist, [track]);
-  }
-
-  /// Adds tracks to a playlist.
-  Future<String?> addTracksToPlaylist(
-    Playlist playlist,
-    List<MediaItem> tracks,
-  ) async {
-    if (_session == null || _offlineMode || tracks.isEmpty) {
-      return 'Playlists are unavailable offline.';
-    }
-    try {
-      await _client.addToPlaylist(
-        playlistId: playlist.id,
-        itemIds: tracks.map((track) => track.id).toList(),
-      );
-      if (_selectedPlaylist?.id == playlist.id) {
-        final refreshed = await _client.fetchPlaylistTracks(playlist.id);
-        _playlistTracks = refreshed;
-        await _cacheStore.savePlaylistTracks(playlist.id, refreshed);
-        notifyListeners();
-      }
-      _updatePlaylistTrackCount(playlist, tracks.length);
-      return null;
-    } catch (error) {
-      return _requestErrorMessage(
-        error,
-        fallback: 'Unable to add to playlist.',
-      );
-    }
-  }
-
-  /// Removes a track from a playlist.
-  Future<String?> removeTrackFromPlaylist(
-    MediaItem track,
-    Playlist playlist,
-  ) async {
-    if (_session == null || _offlineMode) {
-      return 'Playlists are unavailable offline.';
-    }
-    try {
-      await _client.removeFromPlaylist(
-        playlistId: playlist.id,
-        entryIds:
-            track.playlistItemId == null ? const [] : [track.playlistItemId!],
-        itemIds: track.playlistItemId == null ? [track.id] : const [],
-      );
-      if (_selectedPlaylist?.id == playlist.id) {
-        final updated = List<MediaItem>.from(_playlistTracks);
-        if (track.playlistItemId != null) {
-          updated.removeWhere(
-            (item) => item.playlistItemId == track.playlistItemId,
-          );
-        } else {
-          final index = updated.indexWhere((item) => item.id == track.id);
-          if (index != -1) {
-            updated.removeAt(index);
-          }
-        }
-        _playlistTracks = updated;
-        await _cacheStore.savePlaylistTracks(playlist.id, updated);
-        notifyListeners();
-      }
-      _updatePlaylistTrackCount(playlist, -1);
-      return null;
-    } catch (error) {
-      return _requestErrorMessage(
-        error,
-        fallback: 'Unable to remove from playlist.',
-      );
-    }
-  }
-
-  /// Reorders tracks within a playlist.
-  Future<String?> reorderPlaylistTracks(
-    Playlist playlist,
-    List<MediaItem> orderedTracks,
-  ) async {
-    if (_session == null || _offlineMode) {
-      return 'Playlists are unavailable offline.';
-    }
-    final entryIds =
-        orderedTracks.map((track) => track.playlistItemId).toList();
-    if (entryIds.any((id) => id == null)) {
-      return 'Unable to reorder this playlist.';
-    }
-    final previous = _playlistTracks;
-    _playlistTracks = orderedTracks;
-    await _cacheStore.savePlaylistTracks(playlist.id, orderedTracks);
-    notifyListeners();
-    try {
-      await _client.reorderPlaylist(
-        playlistId: playlist.id,
-        entryIds: entryIds.whereType<String>().toList(),
-      );
-      return null;
-    } catch (error) {
-      final fallback = await _attemptPlaylistRebuildReorder(
-        playlist,
-        orderedTracks,
-        error,
-      );
-      if (fallback == null) {
-        return null;
-      }
-      _playlistTracks = previous;
-      await _cacheStore.savePlaylistTracks(playlist.id, previous);
-      notifyListeners();
-      return fallback;
-    }
-  }
-
-  /// Navigates to a library view.
-  void selectLibraryView(LibraryView view, {bool recordHistory = true}) {
-    if (recordHistory && view != _selectedView) {
-      _recordViewHistory(_selectedView);
-    }
-    _selectedView = view;
-    _selectedPlaylist = null;
-    _playlistTracks = [];
-    _selectedSmartList = null;
-    _smartListTracks = [];
-    if (!_isOfflineLibraryView(view) && !_offlineMode) {
-      _offlineOnlyFilter = false;
-    }
-    clearBrowseSelection(notify: false);
-    if (view != LibraryView.home) {
-      clearSearch(notify: false);
-    }
-    notifyListeners();
-    if (view == LibraryView.albums) {
-      unawaited(loadAlbums());
-    }
-    if (view == LibraryView.artists) {
-      unawaited(loadArtists());
-    }
-    if (view == LibraryView.genres) {
-      unawaited(loadGenres());
-    }
-    if (view == LibraryView.tracks) {
-      unawaited(loadLibraryTracks());
-    }
-    if (view == LibraryView.favoritesAlbums) {
-      unawaited(loadFavoriteAlbums());
-    }
-    if (view == LibraryView.favoritesArtists) {
-      unawaited(loadFavoriteArtists());
-    }
-    if (view == LibraryView.favoritesSongs) {
-      unawaited(loadFavoriteTracks());
-    }
-  }
-
-  /// Navigates back to the previous library view.
-  void goBack() {
-    // If in search mode, exit search first
-    if (_isSearching) {
-      clearSearch();
-      return;
-    }
-
-    if (_viewHistory.isEmpty) {
-      return;
-    }
-    final previous = _viewHistory.removeLast();
-    selectLibraryView(previous, recordHistory: false);
   }
 
   void _recordViewHistory(LibraryView view) {
@@ -1659,1144 +963,12 @@ class AppState extends ChangeNotifier {
         view == LibraryView.offlineTracks;
   }
 
-  /// Navigates to an album by identifier.
-  Future<void> selectAlbumById(String albumId) async {
-    if (_albums.isEmpty) {
-      await loadAlbums();
-    }
-    Album? match;
-    for (final album in _albums) {
-      if (album.id == albumId) {
-        match = album;
-        break;
-      }
-    }
-    if (match != null) {
-      await selectAlbum(match, offlineOnly: offlineOnlyFilter);
-    }
-  }
-
-  /// Navigates to an artist by identifier.
-  Future<void> selectArtistById(String artistId) async {
-    if (_artists.isEmpty) {
-      await loadArtists();
-    }
-    Artist? match;
-    for (final artist in _artists) {
-      if (artist.id == artistId) {
-        match = artist;
-        break;
-      }
-    }
-    if (match != null) {
-      await selectArtist(match, offlineOnly: offlineOnlyFilter);
-    }
-  }
-
-  /// Navigates to an artist by name.
-  Future<void> selectArtistByName(String name) async {
-    final trimmed = name.trim();
-    if (trimmed.isEmpty) {
+  void _refreshSelectedSmartList() {
+    final selected = _selectedSmartList;
+    if (selected == null) {
       return;
     }
-    if (_artists.isEmpty) {
-      await loadArtists();
-    }
-    final target = trimmed.toLowerCase();
-    Artist? match;
-    for (final artist in _artists) {
-      if (artist.name.trim().toLowerCase() == target) {
-        match = artist;
-        break;
-      }
-    }
-    if (match == null) {
-      return;
-    }
-    await selectArtist(match, offlineOnly: offlineOnlyFilter);
-  }
-
-  /// Performs a search across the library.
-  Future<void> searchLibrary(String query) async {
-    final trimmed = query.trim();
-    final requestId = ++_searchRequestId;
-
-    // Always enter search mode when this is called
-    if (!_isSearching) {
-      _isSearching = true;
-    }
-
-    if (trimmed.isEmpty) {
-      // Clear results but stay in search view
-      _searchQuery = '';
-      _searchResults = null;
-      _isSearchLoading = false;
-      notifyListeners();
-      return;
-    }
-    _searchQuery = query;
-    _isSearchLoading = true;
-    notifyListeners();
-
-    // Use local search if offline OR if prefer local search is enabled
-    if (_offlineMode || _preferLocalSearch) {
-      await LogService.instance.then((log) => log.info(
-          'Search: Local search for: "$trimmed" (offline=$_offlineMode, preferLocal=$_preferLocalSearch)'));
-
-      _publishLocalSearchResults(trimmed, requestId);
-      await _ensureLocalSearchSourceLoaded(trimmed, requestId);
-      if (!_isSearchRequestActive(requestId, trimmed)) {
-        return;
-      }
-
-      await LogService.instance.then((log) => log.info(
-          'Search: Local results - ${_searchResults?.tracks.length ?? 0} tracks, '
-          '${_searchResults?.albums.length ?? 0} albums, '
-          '${_searchResults?.artists.length ?? 0} artists, '
-          '${_searchResults?.genres.length ?? 0} genres, '
-          '${_searchResults?.playlists.length ?? 0} playlists'));
-
-      _isSearchLoading = false;
-      notifyListeners();
-      return;
-    }
-
-    await LogService.instance.then(
-        (log) => log.info('Search: User initiated search for: "$trimmed"'));
-
-    try {
-      final results = await _client.searchLibrary(trimmed);
-      if (!_isSearchRequestActive(requestId, trimmed)) {
-        return;
-      }
-      _searchResults = results;
-
-      await LogService.instance.then((log) => log.info(
-          'Search: Completed successfully, isEmpty=${_searchResults?.isEmpty}'));
-    } catch (e, stackTrace) {
-      if (!_isSearchRequestActive(requestId, trimmed)) {
-        return;
-      }
-      await LogService.instance
-          .then((log) => log.error('Search: Failed', e, stackTrace));
-      _searchResults = const SearchResults();
-    } finally {
-      if (_isSearchRequestActive(requestId, trimmed)) {
-        _isSearchLoading = false;
-        notifyListeners();
-      }
-    }
-  }
-
-  bool _isSearchRequestActive(int requestId, String trimmedQuery) {
-    if (requestId != _searchRequestId) {
-      return false;
-    }
-    if (!_isSearching) {
-      return false;
-    }
-    return _searchQuery.trim() == trimmedQuery;
-  }
-
-  List<MediaItem> _localSearchTrackSource() {
-    if (!_offlineMode && _libraryTracksFromOfflineSnapshot) {
-      return const <MediaItem>[];
-    }
-    return _libraryTracks;
-  }
-
-  void _publishLocalSearchResults(String query, int requestId) {
-    if (!_isSearchRequestActive(requestId, query)) {
-      return;
-    }
-
-    _searchResults = SearchService.searchLocal(
-      query: query,
-      allTracks: _localSearchTrackSource(),
-      albums: _albums,
-      artists: _artists,
-      genres: _genres,
-      playlists: _playlists,
-    );
-    notifyListeners();
-  }
-
-  Future<void> _ensureLocalSearchSourceLoaded(
-    String query,
-    int requestId,
-  ) async {
-    if (_offlineMode || _session == null) {
-      return;
-    }
-    if (!_isSearchRequestActive(requestId, query)) {
-      return;
-    }
-
-    if (_albums.isEmpty) {
-      await loadAlbums();
-      if (!_isSearchRequestActive(requestId, query)) {
-        return;
-      }
-      _publishLocalSearchResults(query, requestId);
-    }
-    if (_artists.isEmpty) {
-      await loadArtists();
-      if (!_isSearchRequestActive(requestId, query)) {
-        return;
-      }
-      _publishLocalSearchResults(query, requestId);
-    }
-    if (_genres.isEmpty) {
-      await loadGenres();
-      if (!_isSearchRequestActive(requestId, query)) {
-        return;
-      }
-      _publishLocalSearchResults(query, requestId);
-    }
-    if (_playlists.isEmpty) {
-      try {
-        final playlists = await _client.fetchPlaylists();
-        _playlists = playlists;
-        await _cacheStore.savePlaylists(playlists);
-        if (!_isSearchRequestActive(requestId, query)) {
-          return;
-        }
-        _publishLocalSearchResults(query, requestId);
-      } catch (_) {
-        // Keep cached playlists if refresh fails.
-      }
-    }
-
-    if (_libraryTracksFromOfflineSnapshot || _libraryTracks.isEmpty) {
-      await loadLibraryTracks(reset: true);
-      if (!_isSearchRequestActive(requestId, query)) {
-        return;
-      }
-      _publishLocalSearchResults(query, requestId);
-    }
-
-    while (_hasMoreTracks && _isSearchRequestActive(requestId, query)) {
-      final beforeOffset = _tracksOffset;
-      final beforeCount = _libraryTracks.length;
-      await loadLibraryTracks();
-      if (!_isSearchRequestActive(requestId, query)) {
-        return;
-      }
-      if (_tracksOffset == beforeOffset &&
-          _libraryTracks.length == beforeCount) {
-        break;
-      }
-      _publishLocalSearchResults(query, requestId);
-    }
-  }
-
-  /// Updates the search query without triggering a network request.
-  void setSearchQuery(String query, {bool notify = true}) {
-    _searchQuery = query;
-    if (notify) {
-      notifyListeners();
-    }
-  }
-
-  /// Clears the current search results.
-  void clearSearch({bool notify = true}) {
-    _searchQuery = '';
-    _searchResults = null;
-    _isSearching = false;
-    _isSearchLoading = false;
-    if (notify) {
-      notifyListeners();
-    }
-  }
-
-  /// Cycles between repeat off, repeat all, and repeat one.
-  Future<void> toggleRepeatMode() async {
-    switch (_repeatMode) {
-      case LoopMode.off:
-        _repeatMode = LoopMode.all;
-        break;
-      case LoopMode.all:
-        _repeatMode = LoopMode.one;
-        break;
-      case LoopMode.one:
-        _repeatMode = LoopMode.off;
-        break;
-    }
-    await _playback.setLoopMode(_repeatMode);
-    notifyListeners();
-  }
-
-  /// Updates the track browse letter highlight.
-  void setTrackBrowseLetter(String? letter) {
-    if (_trackBrowseLetter == letter) {
-      return;
-    }
-    _trackBrowseLetter = letter;
-    notifyListeners();
-  }
-
-  /// Requests focus for the search field.
-  void requestSearchFocus() {
-    if (_searchQuery.isEmpty && !_isSearching) {
-      // Enter the search view without needing a query.
-      _isSearching = true;
-    }
-    _searchFocusRequest += 1;
-    notifyListeners();
-  }
-
-  /// Loads albums, using cached results when possible.
-  Future<void> loadAlbums() async {
-    final cached = await _cacheStore.loadAlbums();
-    if (cached.isNotEmpty) {
-      _albums = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      _albums = await loadOfflineAlbums();
-      notifyListeners();
-      return;
-    }
-    await _loadAlbums();
-  }
-
-  /// Loads artists, using cached results when possible.
-  Future<void> loadArtists() async {
-    final cached = await _cacheStore.loadArtists();
-    if (cached.isNotEmpty) {
-      _artists = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      _artists = await loadOfflineArtists();
-      notifyListeners();
-      return;
-    }
-    await _loadArtists();
-  }
-
-  /// Loads genres, using cached results when possible.
-  Future<void> loadGenres() async {
-    final cached = await _cacheStore.loadGenres();
-    if (cached.isNotEmpty) {
-      _genres = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      return;
-    }
-    await _loadGenres();
-  }
-
-  /// Loads paginated tracks for the library browse view.
-  Future<void> loadLibraryTracks({bool reset = false}) async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      final offlineTracks = await loadOfflineTracks();
-      _libraryTracks = offlineTracks;
-      _tracksOffset = offlineTracks.length;
-      _hasMoreTracks = false;
-      _isLoadingTracks = false;
-      _libraryTracksFromOfflineSnapshot = true;
-      notifyListeners();
-      return;
-    }
-    if (_isLoadingTracks) {
-      await _tracksLoadCompleter?.future;
-      return;
-    }
-    if (!reset && !_hasMoreTracks) {
-      return;
-    }
-    if (reset) {
-      _libraryTracks = [];
-      _tracksOffset = 0;
-      _hasMoreTracks = true;
-      _libraryTracksFromOfflineSnapshot = false;
-      notifyListeners();
-    }
-    _isLoadingTracks = true;
-    _tracksLoadCompleter = Completer<void>();
-    notifyListeners();
-    try {
-      final tracks = await _client.fetchLibraryTracks(
-        startIndex: _tracksOffset,
-        limit: _tracksPageSize,
-      );
-      if (reset) {
-        _libraryTracks = tracks;
-      } else {
-        _libraryTracks = [..._libraryTracks, ...tracks];
-      }
-      _libraryTracksFromOfflineSnapshot = false;
-      _tracksOffset += tracks.length;
-      if (tracks.length < _tracksPageSize) {
-        _hasMoreTracks = false;
-      }
-    } catch (_) {
-      // Ignore load failures; keep whatever tracks we already have.
-    } finally {
-      _isLoadingTracks = false;
-      _tracksLoadCompleter?.complete();
-      _tracksLoadCompleter = null;
-      notifyListeners();
-    }
-  }
-
-  /// Returns a random track from the library when available.
-  Future<MediaItem?> getRandomTrack() async {
-    if (_offlineMode) {
-      return _randomFromList(_libraryTracks);
-    }
-    if (_session == null) {
-      return null;
-    }
-    try {
-      final track = await _client.fetchRandomTrack();
-      if (track != null) {
-        return track;
-      }
-    } catch (_) {}
-    return _randomFromList(
-          _featuredTracks.isNotEmpty ? _featuredTracks : _recentTracks,
-        ) ??
-        _randomFromList(_libraryTracks);
-  }
-
-  /// Returns a random album from the library when available.
-  Future<Album?> getRandomAlbum() async {
-    if (_offlineMode) {
-      return _randomFromList(_albums);
-    }
-    if (_session == null) {
-      return null;
-    }
-    try {
-      final album = await _client.fetchRandomAlbum();
-      if (album != null) {
-        return album;
-      }
-    } catch (_) {}
-    return _randomFromList(_albums.isNotEmpty ? _albums : _favoriteAlbums);
-  }
-
-  /// Returns a random artist from the library when available.
-  Future<Artist?> getRandomArtist() async {
-    if (_offlineMode) {
-      return _randomFromList(_artists);
-    }
-    if (_session == null) {
-      return null;
-    }
-    try {
-      final artist = await _client.fetchRandomArtist();
-      if (artist != null) {
-        return artist;
-      }
-    } catch (_) {}
-    return _randomFromList(_artists.isNotEmpty ? _artists : _favoriteArtists);
-  }
-
-  /// Loads the Jump in shelf picks.
-  Future<void> loadJumpIn({bool force = false}) async {
-    if (_offlineMode) {
-      if (_isLoadingJumpIn) {
-        return;
-      }
-      _isLoadingJumpIn = true;
-      notifyListeners();
-      _jumpInTrack = _randomFromList(_libraryTracks) ?? _jumpInTrack;
-      _jumpInAlbum = _randomFromList(_albums) ?? _jumpInAlbum;
-      _jumpInArtist = _randomFromList(_artists) ?? _jumpInArtist;
-      _lastJumpInRefreshAt = DateTime.now();
-      _isLoadingJumpIn = false;
-      notifyListeners();
-      return;
-    }
-    if (_session == null) {
-      return;
-    }
-    if (_isLoadingJumpIn) {
-      return;
-    }
-    if (!force &&
-        _jumpInTrack != null &&
-        _jumpInAlbum != null &&
-        _jumpInArtist != null) {
-      return;
-    }
-    _isLoadingJumpIn = true;
-    notifyListeners();
-    try {
-      final results = await Future.wait([
-        getRandomTrack(),
-        getRandomAlbum(),
-        getRandomArtist(),
-      ]);
-      _jumpInTrack = results[0] as MediaItem? ?? _jumpInTrack;
-      _jumpInAlbum = results[1] as Album? ?? _jumpInAlbum;
-      _jumpInArtist = results[2] as Artist? ?? _jumpInArtist;
-      _lastJumpInRefreshAt = DateTime.now();
-    } finally {
-      _isLoadingJumpIn = false;
-      notifyListeners();
-    }
-  }
-
-  /// Plays a shuffled copy of the provided tracks.
-  Future<void> playShuffledList(List<MediaItem> tracks) async {
-    final selection =
-        _offlineMode ? _filterPinnedTracks(tracks) : List.of(tracks);
-    if (selection.isEmpty) {
-      return;
-    }
-    selection.shuffle(_random);
-    await _playFromList(selection, selection.first);
-  }
-
-  /// Loads favorite albums.
-  Future<void> loadFavoriteAlbums() async {
-    final cached = await _cacheStore.loadFavoriteAlbums();
-    if (cached.isNotEmpty) {
-      _favoriteAlbums = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      final offlineAlbums = await loadOfflineAlbums();
-      final offlineIds = offlineAlbums.map((album) => album.id).toSet();
-      _favoriteAlbums =
-          cached.where((album) => offlineIds.contains(album.id)).toList();
-      notifyListeners();
-      return;
-    }
-    await _loadFavoriteAlbums();
-  }
-
-  /// Loads favorite artists.
-  Future<void> loadFavoriteArtists() async {
-    final cached = await _cacheStore.loadFavoriteArtists();
-    if (cached.isNotEmpty) {
-      _favoriteArtists = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      final offlineArtists = await loadOfflineArtists();
-      final offlineIds = offlineArtists.map((artist) => artist.id).toSet();
-      _favoriteArtists =
-          cached.where((artist) => offlineIds.contains(artist.id)).toList();
-      notifyListeners();
-      return;
-    }
-    await _loadFavoriteArtists();
-  }
-
-  /// Loads favorite tracks.
-  Future<void> loadFavoriteTracks() async {
-    final cached = await _cacheStore.loadFavoriteTracks();
-    if (cached.isNotEmpty) {
-      _favoriteTracks = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      _favoriteTracks = _filterPinnedTracks(cached);
-      notifyListeners();
-      return;
-    }
-    await _loadFavoriteTracks();
-  }
-
-  Future<void> _loadSmartListTracks(SmartList list) async {
-    _isLoadingSmartList = true;
-    notifyListeners();
-
-    // Load at least some tracks if we don't have any
-    if (_libraryTracks.isEmpty && !_offlineMode) {
-      await loadLibraryTracks();
-    }
-
-    final results = _buildSmartListTracks(list);
-    _smartListTracks = results;
-    _isLoadingSmartList = false;
-    notifyListeners();
-  }
-
-  Future<void> _ensureSmartListSourceLoaded() async {
-    if (_offlineMode) {
-      if (_libraryTracks.isEmpty) {
-        await loadLibraryTracks();
-      }
-      return;
-    }
-    while (_hasMoreTracks) {
-      await loadLibraryTracks();
-    }
-  }
-
-  List<MediaItem> _buildSmartListTracks(SmartList list) {
-    if (list.scope != SmartListScope.tracks) {
-      return [];
-    }
-    final source = _libraryTracks;
-    final filtered = source
-        .where((track) => _matchesSmartListGroup(list.group, track))
-        .toList();
-    if (list.sorts.isNotEmpty) {
-      filtered.sort((a, b) => _compareSmartListSorts(list.sorts, a, b));
-    }
-    final limit = list.limit;
-    if (limit != null && limit > 0 && filtered.length > limit) {
-      return filtered.take(limit).toList();
-    }
-    return filtered;
-  }
-
-  int _compareSmartListSorts(
-    List<SmartListSort> sorts,
-    MediaItem a,
-    MediaItem b,
-  ) {
-    for (final sort in sorts) {
-      final comparison = _compareSmartListField(sort.field, a, b);
-      if (comparison != 0) {
-        return sort.direction == SmartListSortDirection.desc
-            ? -comparison
-            : comparison;
-      }
-    }
-    return 0;
-  }
-
-  int _compareSmartListField(
-    SmartListField field,
-    MediaItem a,
-    MediaItem b,
-  ) {
-    switch (field) {
-      case SmartListField.title:
-        return a.title.toLowerCase().compareTo(b.title.toLowerCase());
-      case SmartListField.album:
-        return a.album.toLowerCase().compareTo(b.album.toLowerCase());
-      case SmartListField.artist:
-        return a.subtitle.toLowerCase().compareTo(b.subtitle.toLowerCase());
-      case SmartListField.genre:
-        return _joinGenres(a).compareTo(_joinGenres(b));
-      case SmartListField.addedAt:
-        return _compareDate(a.addedAt, b.addedAt);
-      case SmartListField.playCount:
-        return (a.playCount ?? 0).compareTo(b.playCount ?? 0);
-      case SmartListField.lastPlayedAt:
-        return _compareDate(a.lastPlayedAt, b.lastPlayedAt);
-      case SmartListField.duration:
-        return a.duration.compareTo(b.duration);
-      case SmartListField.isFavorite:
-        return _compareBool(isFavoriteTrack(a.id), isFavoriteTrack(b.id));
-      case SmartListField.isDownloaded:
-        return _compareBool(
-          _pinnedAudio.contains(a.streamUrl),
-          _pinnedAudio.contains(b.streamUrl),
-        );
-      case SmartListField.albumIsFavorite:
-        return _compareBool(
-          a.albumId != null && isFavoriteAlbum(a.albumId!),
-          b.albumId != null && isFavoriteAlbum(b.albumId!),
-        );
-      case SmartListField.artistIsFavorite:
-        return _compareBool(
-          a.artistIds.any(isFavoriteArtist),
-          b.artistIds.any(isFavoriteArtist),
-        );
-      case SmartListField.bpm:
-        return (a.bpm ?? 0).compareTo(b.bpm ?? 0);
-    }
-  }
-
-  int _compareBool(bool a, bool b) {
-    if (a == b) {
-      return 0;
-    }
-    return a ? 1 : -1;
-  }
-
-  int _compareDate(DateTime? a, DateTime? b) {
-    if (a == null && b == null) {
-      return 0;
-    }
-    if (a == null) {
-      return -1;
-    }
-    if (b == null) {
-      return 1;
-    }
-    return a.compareTo(b);
-  }
-
-  bool _matchesSmartListGroup(SmartListGroup group, MediaItem track) {
-    if (group.children.isEmpty) {
-      switch (group.mode) {
-        case SmartListGroupMode.all:
-          return true;
-        case SmartListGroupMode.any:
-          return false;
-        case SmartListGroupMode.not:
-          return true;
-      }
-    }
-    final matches = group.children.map((child) {
-      if (child is SmartListRuleNode) {
-        return _matchesSmartListRule(child.rule, track);
-      }
-      if (child is SmartListGroupNode) {
-        return _matchesSmartListGroup(child.group, track);
-      }
-      return false;
-    }).toList();
-    switch (group.mode) {
-      case SmartListGroupMode.all:
-        return matches.every((match) => match);
-      case SmartListGroupMode.any:
-        return matches.any((match) => match);
-      case SmartListGroupMode.not:
-        return !matches.any((match) => match);
-    }
-  }
-
-  bool _matchesSmartListRule(SmartListRule rule, MediaItem track) {
-    switch (rule.field.valueType) {
-      case SmartListValueType.text:
-        final haystack = _valueForTextField(rule.field, track);
-        return _evaluateTextRule(rule, haystack);
-      case SmartListValueType.number:
-        final value = rule.field == SmartListField.playCount
-            ? (track.playCount ?? 0).toDouble()
-            : rule.field == SmartListField.bpm
-                ? (track.bpm ?? 0).toDouble()
-                : 0.0;
-        return _evaluateNumberRule(rule, value);
-      case SmartListValueType.duration:
-        final seconds = track.duration.inSeconds.toDouble();
-        return _evaluateNumberRule(rule, seconds, isDuration: true);
-      case SmartListValueType.date:
-        final date = rule.field == SmartListField.addedAt
-            ? track.addedAt
-            : track.lastPlayedAt;
-        return _evaluateDateRule(rule, date);
-      case SmartListValueType.boolean:
-        final value = _valueForBoolField(rule.field, track);
-        return _evaluateBoolRule(rule, value);
-    }
-  }
-
-  String _valueForTextField(SmartListField field, MediaItem track) {
-    switch (field) {
-      case SmartListField.title:
-        return track.title;
-      case SmartListField.album:
-        return track.album;
-      case SmartListField.artist:
-        return track.subtitle;
-      case SmartListField.genre:
-        return _joinGenres(track);
-      default:
-        return '';
-    }
-  }
-
-  bool _valueForBoolField(SmartListField field, MediaItem track) {
-    switch (field) {
-      case SmartListField.isFavorite:
-        return isFavoriteTrack(track.id);
-      case SmartListField.isDownloaded:
-        return _pinnedAudio.contains(track.streamUrl);
-      case SmartListField.albumIsFavorite:
-        return track.albumId != null && isFavoriteAlbum(track.albumId!);
-      case SmartListField.artistIsFavorite:
-        return track.artistIds.any(isFavoriteArtist);
-      default:
-        return false;
-    }
-  }
-
-  String _joinGenres(MediaItem track) {
-    return track.genres.map((genre) => genre.toLowerCase()).join(', ');
-  }
-
-  bool _evaluateTextRule(SmartListRule rule, String rawValue) {
-    final value = rawValue.toLowerCase();
-    final needle = rule.value.toLowerCase().trim();
-    switch (rule.operatorType) {
-      case SmartListOperator.contains:
-        return needle.isNotEmpty && value.contains(needle);
-      case SmartListOperator.doesNotContain:
-        return needle.isNotEmpty ? !value.contains(needle) : true;
-      case SmartListOperator.equals:
-        return needle.isNotEmpty && value == needle;
-      case SmartListOperator.notEquals:
-        return needle.isNotEmpty ? value != needle : true;
-      case SmartListOperator.startsWith:
-        return needle.isNotEmpty && value.startsWith(needle);
-      case SmartListOperator.endsWith:
-        return needle.isNotEmpty && value.endsWith(needle);
-      default:
-        return false;
-    }
-  }
-
-  bool _evaluateNumberRule(
-    SmartListRule rule,
-    double actualValue, {
-    bool isDuration = false,
-  }) {
-    final value = _parseNumber(rule.value, isDuration: isDuration);
-    final value2 = _parseNumber(rule.value2, isDuration: isDuration);
-    if (value == null) {
-      return false;
-    }
-    switch (rule.operatorType) {
-      case SmartListOperator.equals:
-        return actualValue == value;
-      case SmartListOperator.notEquals:
-        return actualValue != value;
-      case SmartListOperator.greaterThan:
-        return actualValue > value;
-      case SmartListOperator.greaterThanOrEqual:
-        return actualValue >= value;
-      case SmartListOperator.lessThan:
-        return actualValue < value;
-      case SmartListOperator.lessThanOrEqual:
-        return actualValue <= value;
-      case SmartListOperator.between:
-        if (value2 == null) {
-          return false;
-        }
-        final min = value < value2 ? value : value2;
-        final max = value > value2 ? value : value2;
-        return actualValue >= min && actualValue <= max;
-      default:
-        return false;
-    }
-  }
-
-  bool _evaluateDateRule(SmartListRule rule, DateTime? actual) {
-    if (actual == null) {
-      if (rule.operatorType == SmartListOperator.notInLast) {
-        return true;
-      }
-      return false;
-    }
-    switch (rule.operatorType) {
-      case SmartListOperator.isBefore:
-        final target = _parseDate(rule.value);
-        return target != null && actual.isBefore(target);
-      case SmartListOperator.isAfter:
-        final target = _parseDate(rule.value);
-        return target != null && actual.isAfter(target);
-      case SmartListOperator.isOn:
-        final target = _parseDate(rule.value);
-        if (target == null) {
-          return false;
-        }
-        return _isSameDate(actual, target);
-      case SmartListOperator.inLast:
-        final delta = _parseRelativeDuration(rule.value);
-        if (delta == null) {
-          return false;
-        }
-        return actual.isAfter(DateTime.now().subtract(delta));
-      case SmartListOperator.notInLast:
-        final delta = _parseRelativeDuration(rule.value);
-        if (delta == null) {
-          return false;
-        }
-        return actual.isBefore(DateTime.now().subtract(delta));
-      default:
-        return false;
-    }
-  }
-
-  bool _evaluateBoolRule(SmartListRule rule, bool actual) {
-    switch (rule.operatorType) {
-      case SmartListOperator.isTrue:
-        return actual;
-      case SmartListOperator.isFalse:
-        return !actual;
-      default:
-        return false;
-    }
-  }
-
-  double? _parseNumber(String? input, {bool isDuration = false}) {
-    if (input == null) {
-      return null;
-    }
-    final trimmed = input.trim();
-    if (trimmed.isEmpty) {
-      return null;
-    }
-    if (isDuration) {
-      final duration = _parseDuration(trimmed);
-      if (duration != null) {
-        return duration.inSeconds.toDouble();
-      }
-    }
-    return double.tryParse(trimmed);
-  }
-
-  Duration? _parseDuration(String input) {
-    if (input.contains(':')) {
-      final parts = input.split(':').map((part) => part.trim()).toList();
-      if (parts.any((part) => part.isEmpty)) {
-        return null;
-      }
-      final numbers = parts.map(int.tryParse).toList();
-      if (numbers.any((value) => value == null)) {
-        return null;
-      }
-      if (numbers.length == 2) {
-        return Duration(
-          minutes: numbers[0]!,
-          seconds: numbers[1]!,
-        );
-      }
-      if (numbers.length == 3) {
-        return Duration(
-          hours: numbers[0]!,
-          minutes: numbers[1]!,
-          seconds: numbers[2]!,
-        );
-      }
-    }
-    final seconds = int.tryParse(input);
-    if (seconds == null) {
-      return null;
-    }
-    return Duration(seconds: seconds);
-  }
-
-  DateTime? _parseDate(String input) {
-    final trimmed = input.trim();
-    if (trimmed.isEmpty) {
-      return null;
-    }
-    return DateTime.tryParse(trimmed);
-  }
-
-  bool _isSameDate(DateTime a, DateTime b) {
-    return a.year == b.year && a.month == b.month && a.day == b.day;
-  }
-
-  Duration? _parseRelativeDuration(String input) {
-    final trimmed = input.trim().toLowerCase();
-    final match = RegExp(r'^(\d+)([dwmy])$').firstMatch(trimmed);
-    if (match == null) {
-      return null;
-    }
-    final amount = int.tryParse(match.group(1) ?? '');
-    final unit = match.group(2);
-    if (amount == null || amount <= 0 || unit == null) {
-      return null;
-    }
-    switch (unit) {
-      case 'd':
-        return Duration(days: amount);
-      case 'w':
-        return Duration(days: amount * 7);
-      case 'm':
-        return Duration(days: amount * 30);
-      case 'y':
-        return Duration(days: amount * 365);
-      default:
-        return null;
-    }
-  }
-
-  /// Selects an album and loads its tracks.
-  Future<void> selectAlbum(Album album, {bool offlineOnly = false}) async {
-    if (_selectedView != LibraryView.home) {
-      _recordViewHistory(_selectedView);
-    }
-    _selectedAlbum = album;
-    _selectedSmartList = null;
-    _selectedArtist = null;
-    _selectedGenre = null;
-    _offlineOnlyFilter = offlineOnly;
-    clearSearch(notify: false);
-    notifyListeners();
-    final cached = await _cacheStore.loadAlbumTracks(album.id);
-    if (cached.isNotEmpty) {
-      _albumTracks = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      final filtered = _filterPinnedTracks(_albumTracks);
-      _albumTracks =
-          filtered.isNotEmpty ? filtered : await _offlineTracksForAlbum(album);
-      notifyListeners();
-      return;
-    }
-    try {
-      final tracks = await _client.fetchAlbumTracks(album.id);
-      _albumTracks = tracks;
-      await _cacheStore.saveAlbumTracks(album.id, tracks);
-      notifyListeners();
-    } catch (_) {
-      // Keep cached tracks if refresh fails.
-    }
-  }
-
-  /// Loads an album and starts playback.
-  Future<void> playAlbum(Album album) async {
-    final logService = await LogService.instance;
-    await logService.info(
-        'playAlbum: Starting "${album.name}" (${album.id}), offline=$_offlineMode');
-
-    await selectAlbum(album);
-    final tracks =
-        _offlineMode ? _filterPinnedTracks(_albumTracks) : _albumTracks;
-    if (tracks.isNotEmpty) {
-      await logService.info('playAlbum: Playing ${tracks.length} tracks');
-      await _playFromList(tracks, tracks.first);
-    } else {
-      await logService.warning('playAlbum: No tracks available');
-    }
-  }
-
-  /// Selects an artist and loads their tracks.
-  Future<void> selectArtist(Artist artist, {bool offlineOnly = false}) async {
-    if (_selectedView != LibraryView.home) {
-      _recordViewHistory(_selectedView);
-    }
-    _selectedArtist = artist;
-    _selectedSmartList = null;
-    _selectedAlbum = null;
-    _selectedGenre = null;
-    _offlineOnlyFilter = offlineOnly;
-    clearSearch(notify: false);
-    notifyListeners();
-    final cached = await _cacheStore.loadArtistTracks(artist.id);
-    if (cached.isNotEmpty) {
-      _artistTracks = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      final filtered = _filterPinnedTracks(_artistTracks);
-      _artistTracks = filtered.isNotEmpty
-          ? filtered
-          : await _offlineTracksForArtist(artist);
-      notifyListeners();
-      return;
-    }
-    try {
-      final tracks = await _client.fetchArtistTracks(artist.id);
-      _artistTracks = tracks;
-      await _cacheStore.saveArtistTracks(artist.id, tracks);
-      notifyListeners();
-    } catch (_) {
-      // Keep cached tracks if refresh fails.
-    }
-  }
-
-  /// Loads an artist and starts playback.
-  Future<void> playArtist(Artist artist) async {
-    await selectArtist(artist);
-    final tracks =
-        _offlineMode ? _filterPinnedTracks(_artistTracks) : _artistTracks;
-    if (tracks.isNotEmpty) {
-      await _playFromList(tracks, tracks.first);
-    }
-  }
-
-  /// Selects a genre and loads its tracks.
-  Future<void> selectGenre(Genre genre) async {
-    if (_selectedView != LibraryView.home) {
-      _recordViewHistory(_selectedView);
-    }
-    _selectedGenre = genre;
-    _selectedSmartList = null;
-    _selectedAlbum = null;
-    _selectedArtist = null;
-    clearSearch(notify: false);
-    notifyListeners();
-    final cached = await _cacheStore.loadGenreTracks(genre.id);
-    if (cached.isNotEmpty) {
-      _genreTracks = cached;
-      notifyListeners();
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      final tracks = await _client.fetchGenreTracks(genre.id);
-      _genreTracks = tracks;
-      await _cacheStore.saveGenreTracks(genre.id, tracks);
-      notifyListeners();
-    } catch (_) {
-      // Keep cached tracks if refresh fails.
-    }
-  }
-
-  /// Loads a genre and starts playback.
-  Future<void> playGenre(Genre genre) async {
-    await selectGenre(genre);
-    final tracks =
-        _offlineMode ? _filterPinnedTracks(_genreTracks) : _genreTracks;
-    if (tracks.isNotEmpty) {
-      await _playFromList(tracks, tracks.first);
-    }
-  }
-
-  /// Starts playback from a selected track.
-  Future<void> playFromPlaylist(MediaItem track) async {
-    await playFromList(_playlistTracks, track);
-  }
-
-  /// Plays tracks from the selected album.
-  Future<void> playFromAlbum(MediaItem track) async {
-    await playFromList(_albumTracks, track);
-  }
-
-  /// Plays tracks from the selected artist.
-  Future<void> playFromArtist(MediaItem track) async {
-    await playFromList(_artistTracks, track);
-  }
-
-  /// Plays tracks from the selected genre.
-  Future<void> playFromGenre(MediaItem track) async {
-    await playFromList(_genreTracks, track);
-  }
-
-  /// Plays tracks from favorites.
-  Future<void> playFromFavorites(MediaItem track) async {
-    await playFromList(_favoriteTracks, track);
-  }
-
-  /// Plays tracks from search results.
-  Future<void> playFromSearch(MediaItem track) async {
-    final tracks = _searchResults?.tracks ?? const <MediaItem>[];
-    await playFromList(tracks, track);
-  }
-
-  /// Plays tracks from a provided list.
-  Future<void> playFromList(List<MediaItem> tracks, MediaItem track) async {
-    if (_offlineMode) {
-      final filtered = _filterPinnedTracks(tracks);
-      if (filtered.isEmpty) {
-        return;
-      }
-      final match = filtered.firstWhere(
-        (item) => item.id == track.id,
-        orElse: () => filtered.first,
-      );
-      await _playFromList(filtered, match);
-      return;
-    }
-    await _playFromList(tracks, track);
-  }
-
-  /// Plays featured tracks from the home shelf.
-  Future<void> playFeatured(MediaItem track) async {
-    await playFromList(_featuredTracks, track);
+    unawaited(_loadSmartListTracks(selected));
   }
 
   /// Toggles between play and pause states.
@@ -2991,994 +1163,6 @@ class AppState extends ChangeNotifier {
     }
   }
 
-  /// Updates the theme preference.
-  Future<void> setThemeMode(ThemeMode mode) async {
-    _themeMode = mode;
-    await _settingsStore.saveThemeMode(mode);
-    notifyListeners();
-  }
-
-  /// Updates playback telemetry preference.
-  Future<void> setTelemetryPlayback(bool enabled) async {
-    _telemetryPlayback = enabled;
-    await _settingsStore.savePlaybackTelemetry(enabled);
-    notifyListeners();
-  }
-
-  /// Updates playback progress telemetry preference.
-  Future<void> setTelemetryProgress(bool enabled) async {
-    _telemetryProgress = enabled;
-    await _settingsStore.saveProgressTelemetry(enabled);
-    notifyListeners();
-  }
-
-  /// Updates playback history telemetry preference.
-  Future<void> setTelemetryHistory(bool enabled) async {
-    _telemetryHistory = enabled;
-    await _settingsStore.saveHistoryTelemetry(enabled);
-    notifyListeners();
-  }
-
-  /// Updates the gapless playback preference.
-  Future<void> setGaplessPlayback(bool enabled) async {
-    if (_gaplessPlayback == enabled) {
-      return;
-    }
-    _gaplessPlayback = enabled;
-    await _settingsStore.saveGaplessPlayback(enabled);
-    await _applyPlaybackSettings();
-    notifyListeners();
-  }
-
-  /// Updates whether downloads are limited to Wi-Fi.
-  Future<void> setDownloadsWifiOnly(bool enabled) async {
-    _downloadsWifiOnly = enabled;
-    await _settingsStore.saveDownloadsWifiOnly(enabled);
-    notifyListeners();
-    unawaited(_processDownloadQueue());
-  }
-
-  /// Updates whether downloads are paused.
-  Future<void> setDownloadsPaused(bool paused) async {
-    _downloadsPaused = paused;
-    await _settingsStore.saveDownloadsPaused(paused);
-    notifyListeners();
-    if (!paused) {
-      unawaited(_processDownloadQueue());
-    }
-  }
-
-  /// Updates auto-download preference for favorites.
-  Future<void> setAutoDownloadFavoritesEnabled(bool enabled) async {
-    _autoDownloadFavoritesEnabled = enabled;
-    await _settingsStore.saveAutoDownloadFavoritesEnabled(enabled);
-    notifyListeners();
-    if (enabled) {
-      unawaited(_prefetchFavoriteDownloads());
-    }
-  }
-
-  /// Updates auto-download preference for favorited albums.
-  Future<void> setAutoDownloadFavoriteAlbums(bool enabled) async {
-    _autoDownloadFavoriteAlbums = enabled;
-    await _settingsStore.saveAutoDownloadFavoriteAlbums(enabled);
-    notifyListeners();
-    if (enabled && _autoDownloadFavoritesEnabled) {
-      unawaited(_prefetchFavoriteDownloads(albumsOnly: true));
-    }
-  }
-
-  /// Updates auto-download preference for favorited artists.
-  Future<void> setAutoDownloadFavoriteArtists(bool enabled) async {
-    _autoDownloadFavoriteArtists = enabled;
-    await _settingsStore.saveAutoDownloadFavoriteArtists(enabled);
-    notifyListeners();
-    if (enabled && _autoDownloadFavoritesEnabled) {
-      unawaited(_prefetchFavoriteDownloads(artistsOnly: true));
-    }
-  }
-
-  /// Updates auto-download preference for favorited tracks.
-  Future<void> setAutoDownloadFavoriteTracks(bool enabled) async {
-    _autoDownloadFavoriteTracks = enabled;
-    await _settingsStore.saveAutoDownloadFavoriteTracks(enabled);
-    notifyListeners();
-    if (enabled && _autoDownloadFavoritesEnabled) {
-      unawaited(_prefetchFavoriteDownloads(tracksOnly: true));
-    }
-  }
-
-  /// Updates Wi-Fi only auto-download preference.
-  Future<void> setAutoDownloadFavoritesWifiOnly(bool enabled) async {
-    _autoDownloadFavoritesWifiOnly = enabled;
-    await _settingsStore.saveAutoDownloadFavoritesWifiOnly(enabled);
-    notifyListeners();
-    if (enabled && _autoDownloadFavoritesEnabled) {
-      unawaited(_prefetchFavoriteDownloads());
-    }
-  }
-
-  /// Updates the ordering of pending downloads.
-  void reorderDownloadQueue(int oldIndex, int newIndex) {
-    if (oldIndex == newIndex) {
-      return;
-    }
-    if (oldIndex < 0 || oldIndex >= _downloadQueue.length) {
-      return;
-    }
-    if (newIndex > oldIndex) {
-      newIndex -= 1;
-    }
-    final task = _downloadQueue.removeAt(oldIndex);
-    final target = newIndex.clamp(0, _downloadQueue.length);
-    _downloadQueue.insert(target, task);
-    notifyListeners();
-  }
-
-  /// Retries a failed download.
-  void retryDownload(DownloadTask task) {
-    final index = _indexOfDownload(task.track.streamUrl);
-    if (index == null) {
-      return;
-    }
-    final queuedTask = _downloadQueue[index].copyWith(
-      status: DownloadStatus.queued,
-      progress: null,
-      totalBytes: null,
-      downloadedBytes: null,
-      errorMessage: null,
-    );
-    _replaceDownloadTaskAt(index, queuedTask);
-    notifyListeners();
-    unawaited(_processDownloadQueue());
-  }
-
-  /// Updates the offline mode preference.
-  Future<void> setOfflineMode(bool enabled) async {
-    if (_offlineMode == enabled) {
-      return;
-    }
-    _offlineMode = enabled;
-    await _settingsStore.saveOfflineMode(enabled);
-    if (enabled) {
-      await _applyOfflineModeData();
-      return;
-    }
-    _offlineOnlyFilter = false;
-    _libraryTracks = [];
-    _tracksOffset = 0;
-    _hasMoreTracks = true;
-    _isLoadingTracks = false;
-    _tracksLoadCompleter = null;
-    _libraryTracksFromOfflineSnapshot = false;
-    notifyListeners();
-    if (_session != null) {
-      unawaited(refreshLibrary());
-    }
-    unawaited(_processDownloadQueue());
-  }
-
-  /// Updates the offline-only filter for detail views.
-  void setOfflineOnlyFilter(bool enabled) {
-    final next = _offlineMode ? true : enabled;
-    if (_offlineOnlyFilter == next) {
-      return;
-    }
-    _offlineOnlyFilter = next;
-    notifyListeners();
-  }
-
-  /// Updates the font family preference.
-  Future<void> setFontFamily(String? family) async {
-    _fontFamily = family;
-    await _settingsStore.saveFontFamily(family);
-    notifyListeners();
-  }
-
-  /// Updates the font scale preference.
-  Future<void> setFontScale(double scale) async {
-    _fontScale = scale;
-    await _settingsStore.saveFontScale(scale);
-    notifyListeners();
-  }
-
-  /// Updates the accent color preference.
-  Future<void> setAccentColor(Color color) async {
-    _accentColorValue = color.toARGB32();
-    await _settingsStore.saveAccentColorValue(color.toARGB32());
-    notifyListeners();
-  }
-
-  /// Updates the accent color source preference.
-  Future<void> setAccentColorSource(AccentColorSource source) async {
-    _accentColorSource = source;
-    await _settingsStore.saveAccentColorSource(source);
-    notifyListeners();
-    if (source == AccentColorSource.nowPlaying) {
-      unawaited(_maybeUpdateNowPlayingPalette(_nowPlaying));
-    }
-  }
-
-  /// Updates the theme palette source preference.
-  Future<void> setThemePaletteSource(ThemePaletteSource source) async {
-    _themePaletteSource = source;
-    await _settingsStore.saveThemePaletteSource(source);
-    notifyListeners();
-    if (source == ThemePaletteSource.nowPlaying) {
-      unawaited(_maybeUpdateNowPlayingPalette(_nowPlaying));
-    }
-  }
-
-  /// Updates the settings shortcut enabled preference.
-  Future<void> setSettingsShortcutEnabled(bool enabled) async {
-    _settingsShortcutEnabled = enabled;
-    await _settingsStore.saveSettingsShortcutEnabled(enabled);
-    notifyListeners();
-  }
-
-  /// Updates the settings shortcut preference.
-  Future<void> setSettingsShortcut(KeyboardShortcut shortcut) async {
-    _settingsShortcut = shortcut;
-    await _settingsStore.saveSettingsShortcut(shortcut);
-    notifyListeners();
-  }
-
-  /// Updates the search shortcut enabled preference.
-  Future<void> setSearchShortcutEnabled(bool enabled) async {
-    _searchShortcutEnabled = enabled;
-    await _settingsStore.saveSearchShortcutEnabled(enabled);
-    notifyListeners();
-  }
-
-  /// Updates the search shortcut preference.
-  Future<void> setSearchShortcut(KeyboardShortcut shortcut) async {
-    _searchShortcut = shortcut;
-    await _settingsStore.saveSearchShortcut(shortcut);
-    notifyListeners();
-  }
-
-  /// Updates the prefer local search preference.
-  Future<void> setPreferLocalSearch(bool enabled) async {
-    _preferLocalSearch = enabled;
-    await _settingsStore.savePreferLocalSearch(enabled);
-    notifyListeners();
-  }
-
-  /// Updates the layout density preference.
-  Future<void> setLayoutDensity(LayoutDensity density) async {
-    _layoutDensity = density;
-    await _settingsStore.saveLayoutDensity(density);
-    notifyListeners();
-  }
-
-  /// Updates the corner radius style preference.
-  Future<void> setCornerRadiusStyle(CornerRadiusStyle style) async {
-    _cornerRadiusStyle = style;
-    await _settingsStore.saveCornerRadiusStyle(style);
-    notifyListeners();
-  }
-
-  /// Updates the track list style preference.
-  Future<void> setTrackListStyle(TrackListStyle style) async {
-    _trackListStyle = style;
-    await _settingsStore.saveTrackListStyle(style);
-    notifyListeners();
-  }
-
-  /// Updates whether track timestamp status icons are shown.
-  Future<void> setTrackStatusIconsEnabled(bool enabled) async {
-    _trackStatusIconsEnabled = enabled;
-    await _settingsStore.saveTrackStatusIconsEnabled(enabled);
-    notifyListeners();
-  }
-
-  /// Updates the now playing layout preference.
-  Future<void> setNowPlayingLayout(NowPlayingLayout layout) async {
-    _nowPlayingLayout = layout;
-    await _settingsStore.saveNowPlayingLayout(layout);
-    notifyListeners();
-  }
-
-  /// Updates the home shelf layout preference.
-  Future<void> setHomeShelfLayout(HomeShelfLayout layout) async {
-    _homeShelfLayout = layout;
-    await _settingsStore.saveHomeShelfLayout(layout);
-    notifyListeners();
-  }
-
-  /// Updates the home shelf grid row count.
-  Future<void> setHomeShelfGridRows(int rows) async {
-    _homeShelfGridRows = rows;
-    await _settingsStore.saveHomeShelfGridRows(rows);
-    notifyListeners();
-  }
-
-  /// Updates the sidebar width preference.
-  Future<void> setSidebarWidth(
-    double width, {
-    bool persist = true,
-  }) async {
-    _sidebarWidth = width;
-    if (persist) {
-      await _settingsStore.saveSidebarWidth(width);
-    }
-    notifyListeners();
-  }
-
-  /// Updates the sidebar collapsed preference.
-  Future<void> setSidebarCollapsed(
-    bool collapsed, {
-    bool persist = true,
-  }) async {
-    _sidebarCollapsed = collapsed;
-    if (persist) {
-      await _settingsStore.saveSidebarCollapsed(collapsed);
-    }
-    notifyListeners();
-  }
-
-  /// Updates the sidebar overlay open state.
-  void setSidebarOverlayOpen(bool open) {
-    if (_sidebarOverlayOpen == open) {
-      return;
-    }
-    _sidebarOverlayOpen = open;
-    notifyListeners();
-  }
-
-  /// Toggles the sidebar overlay open state.
-  void toggleSidebarOverlayOpen() {
-    setSidebarOverlayOpen(!_sidebarOverlayOpen);
-  }
-
-  /// Clears cached metadata entries.
-  Future<void> clearMetadataCache() async {
-    await _cacheStore.clearMetadata();
-    await refreshLibrary();
-  }
-
-  /// Clears cached audio files.
-  Future<void> clearAudioCache() async {
-    await _cacheStore.clearAudioCache();
-    await refreshMediaCacheBytes();
-  }
-
-  /// Returns the estimated cached media size in bytes.
-  Future<int> getMediaCacheBytes() async {
-    return _cacheStore.getMediaCacheBytes();
-  }
-
-  /// Refreshes cached media size counters.
-  Future<void> refreshMediaCacheBytes() async {
-    final totalBytes = await _cacheStore.getMediaCacheBytes();
-    _mediaCacheBytesNotifier.value = totalBytes;
-    final pinnedBytes = await _cacheStore.getPinnedMediaBytes(_pinnedAudio);
-    _pinnedCacheBytesNotifier.value = pinnedBytes;
-  }
-
-  /// Returns the estimated size of pinned downloads.
-  Future<int> getPinnedCacheBytes() async {
-    return _cacheStore.getPinnedMediaBytes(_pinnedAudio);
-  }
-
-  /// Updates the cache size limit.
-  Future<void> setCacheMaxBytes(int bytes) async {
-    _cacheMaxBytes = bytes;
-    await _cacheStore.saveCacheMaxBytes(bytes);
-    await refreshMediaCacheBytes();
-    notifyListeners();
-  }
-
-  /// Returns cached audio entries for display.
-  Future<List<CachedAudioEntry>> getCachedAudioEntries() async {
-    return _cacheStore.loadCachedAudioEntries();
-  }
-
-  /// Opens the cached media location in the OS file manager.
-  Future<void> showMediaCacheLocation() async {
-    await _cacheStore.openMediaCacheLocation();
-  }
-
-  /// Removes a cached audio entry and its file.
-  Future<void> evictCachedAudio(String streamUrl) async {
-    await _cacheStore.evictCachedAudio(streamUrl);
-    await refreshMediaCacheBytes();
-  }
-
-  Future<bool> _canDownloadOverNetwork({bool requireWifi = false}) async {
-    if (_offlineMode) {
-      return false;
-    }
-    final needsWifi = _downloadsWifiOnly || requireWifi;
-    if (!needsWifi) {
-      return true;
-    }
-    if (kIsWeb) {
-      return true;
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.macOS:
-      case TargetPlatform.windows:
-      case TargetPlatform.linux:
-        return true;
-      case TargetPlatform.android:
-      case TargetPlatform.iOS:
-        try {
-          final statuses = await Connectivity().checkConnectivity();
-          return statuses.any(_networkConnectivityWhitelist.contains);
-        } catch (_) {
-          return false;
-        }
-      case TargetPlatform.fuchsia:
-        return true;
-    }
-  }
-
-  Future<void> _prefetchFavoriteDownloads({
-    bool albumsOnly = false,
-    bool artistsOnly = false,
-    bool tracksOnly = false,
-  }) async {
-    if (!_autoDownloadFavoritesEnabled) {
-      return;
-    }
-    final shouldAlbums =
-        _autoDownloadFavoriteAlbums && !artistsOnly && !tracksOnly;
-    final shouldArtists =
-        _autoDownloadFavoriteArtists && !albumsOnly && !tracksOnly;
-    final shouldTracks =
-        _autoDownloadFavoriteTracks && !albumsOnly && !artistsOnly;
-    if (shouldAlbums) {
-      for (final album in _favoriteAlbums) {
-        await makeAlbumAvailableOffline(
-          album,
-          requiresWifi: _autoDownloadFavoritesWifiOnly,
-        );
-      }
-    }
-    if (shouldArtists) {
-      for (final artist in _favoriteArtists) {
-        await makeArtistAvailableOffline(
-          artist,
-          requiresWifi: _autoDownloadFavoritesWifiOnly,
-        );
-      }
-    }
-    if (shouldTracks) {
-      for (final track in _favoriteTracks) {
-        await makeTrackAvailableOffline(
-          track,
-          requiresWifi: _autoDownloadFavoritesWifiOnly,
-        );
-      }
-    }
-  }
-
-  int? _indexOfDownload(String streamUrl) {
-    final index = _downloadQueue.indexWhere(
-      (task) => task.track.streamUrl == streamUrl,
-    );
-    return index == -1 ? null : index;
-  }
-
-  void _addDownloadTask(DownloadTask task) {
-    _downloadQueue.add(task);
-    _downloadStatusByUrl[task.track.streamUrl] = task.status;
-  }
-
-  void _replaceDownloadTaskAt(int index, DownloadTask task) {
-    _downloadQueue[index] = task;
-    _downloadStatusByUrl[task.track.streamUrl] = task.status;
-  }
-
-  void _removeDownloadTaskAt(int index) {
-    final removed = _downloadQueue.removeAt(index);
-    _downloadStatusByUrl.remove(removed.track.streamUrl);
-  }
-
-  Future<void> _queueDownload(
-    MediaItem track, {
-    bool requiresWifi = false,
-  }) async {
-    final normalized = _normalizeTrackForPlayback(track);
-    _cancelledOfflineRequests.remove(normalized.streamUrl);
-    final existingIndex = _indexOfDownload(normalized.streamUrl);
-    if (existingIndex != null) {
-      final existing = _downloadQueue[existingIndex];
-      if (existing.status == DownloadStatus.failed) {
-        retryDownload(existing);
-      }
-      return;
-    }
-    final cached = await _cacheStore.isAudioCached(normalized);
-    if (cached) {
-      await _cacheStore.touchCachedAudio(normalized);
-      return;
-    }
-    _addDownloadTask(
-      DownloadTask(
-        track: normalized,
-        status: DownloadStatus.queued,
-        queuedAt: DateTime.now(),
-        requiresWifi: requiresWifi,
-      ),
-    );
-    notifyListeners();
-    unawaited(_processDownloadQueue());
-  }
-
-  void _resetWaitingDownloads() {
-    var updated = false;
-    for (var i = 0; i < _downloadQueue.length; i += 1) {
-      final task = _downloadQueue[i];
-      if (task.status == DownloadStatus.waitingForWifi) {
-        _replaceDownloadTaskAt(
-          i,
-          task.copyWith(status: DownloadStatus.queued),
-        );
-        updated = true;
-      }
-    }
-    if (updated) {
-      notifyListeners();
-    }
-  }
-
-  void _updateDownloadTask(
-    String streamUrl, {
-    DownloadStatus? status,
-    double? progress,
-    int? totalBytes,
-    int? downloadedBytes,
-    String? errorMessage,
-  }) {
-    final index = _indexOfDownload(streamUrl);
-    if (index == null) {
-      return;
-    }
-    final existing = _downloadQueue[index];
-    if (progress != null &&
-        _shouldThrottleProgress(
-          streamUrl,
-          progress,
-          existing.progress,
-        )) {
-      return;
-    }
-    final nextTask = existing.copyWith(
-      status: status,
-      progress: progress,
-      totalBytes: totalBytes,
-      downloadedBytes: downloadedBytes,
-      errorMessage: errorMessage,
-    );
-    _replaceDownloadTaskAt(index, nextTask);
-    notifyListeners();
-  }
-
-  bool _shouldThrottleProgress(
-    String streamUrl,
-    double newProgress,
-    double? currentProgress,
-  ) {
-    final now = DateTime.now();
-    final last = _downloadProgressTimestamps[streamUrl];
-    final pivot = currentProgress ?? 0.0;
-    if (last != null &&
-        now.difference(last) < const Duration(milliseconds: 250) &&
-        (newProgress - pivot).abs() < 0.01) {
-      return true;
-    }
-    _downloadProgressTimestamps[streamUrl] = now;
-    return false;
-  }
-
-  void _removeDownload(String streamUrl) {
-    final index = _indexOfDownload(streamUrl);
-    if (index == null) {
-      return;
-    }
-    _removeDownloadTaskAt(index);
-    _downloadProgressTimestamps.remove(streamUrl);
-    notifyListeners();
-  }
-
-  Future<void> _processDownloadQueue() async {
-    if (_isProcessingDownloads) {
-      return;
-    }
-    if (_downloadsPaused) {
-      return;
-    }
-    _isProcessingDownloads = true;
-    try {
-      _resetWaitingDownloads();
-      while (true) {
-        if (_downloadsPaused) {
-          break;
-        }
-        DownloadTask? next;
-        // Iterate over a copy to avoid concurrent modification
-        for (final task in List.from(_downloadQueue)) {
-          if (task.status == DownloadStatus.queued) {
-            final canDownload = await _canDownloadOverNetwork(
-              requireWifi: task.requiresWifi,
-            );
-            if (canDownload) {
-              next = task;
-              break;
-            }
-            _updateDownloadTask(
-              task.track.streamUrl,
-              status: DownloadStatus.waitingForWifi,
-            );
-          }
-        }
-        if (next == null) {
-          break;
-        }
-        await _downloadTrack(next);
-      }
-    } finally {
-      _isProcessingDownloads = false;
-    }
-  }
-
-  Future<void> _downloadTrack(DownloadTask task) async {
-    final streamUrl = task.track.streamUrl;
-    _updateDownloadTask(streamUrl, status: DownloadStatus.downloading);
-    try {
-      await for (final response in _cacheStore.downloadAudioWithProgress(
-        task.track,
-        headers: _playbackHeaders(),
-      )) {
-        if (_cancelledOfflineRequests.contains(streamUrl)) {
-          _cancelledOfflineRequests.remove(streamUrl);
-          _removeDownload(streamUrl);
-          return;
-        }
-        if (response is DownloadProgress) {
-          _updateDownloadTask(
-            streamUrl,
-            progress: response.progress,
-            totalBytes: response.totalSize,
-            downloadedBytes: response.downloaded,
-          );
-        } else if (response is FileInfo) {
-          _removeDownload(streamUrl);
-          unawaited(refreshMediaCacheBytes());
-        }
-      }
-    } catch (error) {
-      if (_cancelledOfflineRequests.contains(streamUrl)) {
-        _cancelledOfflineRequests.remove(streamUrl);
-        _removeDownload(streamUrl);
-        return;
-      }
-      _updateDownloadTask(
-        streamUrl,
-        status: DownloadStatus.failed,
-        errorMessage: error.toString(),
-      );
-    }
-  }
-
-  /// Pins a track for offline playback.
-  Future<void> makeTrackAvailableOffline(
-    MediaItem track, {
-    bool requiresWifi = false,
-  }) async {
-    _cancelledOfflineRequests.remove(track.streamUrl);
-    await _cacheStore.setPinnedAudio(track.streamUrl, true);
-    _pinnedAudio.add(track.streamUrl);
-    await _queueDownload(track, requiresWifi: requiresWifi);
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  /// Removes a track from offline pinning.
-  Future<void> unpinTrackOffline(MediaItem track) async {
-    _cancelledOfflineRequests.add(track.streamUrl);
-    await _cacheStore.setPinnedAudio(track.streamUrl, false);
-    _pinnedAudio.remove(track.streamUrl);
-    _removeDownload(track.streamUrl);
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  Future<List<MediaItem>> _loadAlbumTracksForOffline(Album album) async {
-    final cached = await _cacheStore.loadAlbumTracks(album.id);
-    if (cached.isNotEmpty) {
-      return cached;
-    }
-    if (_offlineMode) {
-      return [];
-    }
-    try {
-      final tracks = await _client.fetchAlbumTracks(album.id);
-      await _cacheStore.saveAlbumTracks(album.id, tracks);
-      return tracks;
-    } catch (_) {
-      return [];
-    }
-  }
-
-  Future<List<MediaItem>> _loadArtistTracksForOffline(Artist artist) async {
-    final cached = await _cacheStore.loadArtistTracks(artist.id);
-    if (cached.isNotEmpty) {
-      return cached;
-    }
-    if (_offlineMode) {
-      return [];
-    }
-    try {
-      final tracks = await _client.fetchArtistTracks(artist.id);
-      await _cacheStore.saveArtistTracks(artist.id, tracks);
-      return tracks;
-    } catch (_) {
-      return [];
-    }
-  }
-
-  Future<List<MediaItem>> _loadPlaylistTracksForOffline(
-      Playlist playlist) async {
-    final cached = await _cacheStore.loadPlaylistTracks(playlist.id);
-    if (cached.isNotEmpty) {
-      return cached;
-    }
-    if (_offlineMode) {
-      return [];
-    }
-    try {
-      final tracks = await _client.fetchPlaylistTracks(playlist.id);
-      await _cacheStore.savePlaylistTracks(playlist.id, tracks);
-      return tracks;
-    } catch (_) {
-      return [];
-    }
-  }
-
-  /// Pins all tracks in a playlist for offline playback.
-  Future<void> makePlaylistAvailableOffline(
-    Playlist playlist, {
-    bool requiresWifi = false,
-  }) async {
-    final tracks = await _loadPlaylistTracksForOffline(playlist);
-    for (final track in tracks) {
-      _cancelledOfflineRequests.remove(track.streamUrl);
-      await _cacheStore.setPinnedAudio(track.streamUrl, true);
-      _pinnedAudio.add(track.streamUrl);
-      await _queueDownload(track, requiresWifi: requiresWifi);
-    }
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  /// Removes a playlist from offline pinning.
-  Future<void> unpinPlaylistOffline(Playlist playlist) async {
-    final tracks = await _loadPlaylistTracksForOffline(playlist);
-    for (final track in tracks) {
-      _cancelledOfflineRequests.add(track.streamUrl);
-      await _cacheStore.setPinnedAudio(track.streamUrl, false);
-      _pinnedAudio.remove(track.streamUrl);
-      _removeDownload(track.streamUrl);
-    }
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  /// Pins all tracks in an album for offline playback.
-  Future<void> makeAlbumAvailableOffline(
-    Album album, {
-    bool requiresWifi = false,
-  }) async {
-    final tracks = await _loadAlbumTracksForOffline(album);
-    for (final track in tracks) {
-      _cancelledOfflineRequests.remove(track.streamUrl);
-      await _cacheStore.setPinnedAudio(track.streamUrl, true);
-      _pinnedAudio.add(track.streamUrl);
-      await _queueDownload(track, requiresWifi: requiresWifi);
-    }
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  /// Removes an album from offline pinning.
-  Future<void> unpinAlbumOffline(Album album) async {
-    final tracks = await _loadAlbumTracksForOffline(album);
-    for (final track in tracks) {
-      _cancelledOfflineRequests.add(track.streamUrl);
-      await _cacheStore.setPinnedAudio(track.streamUrl, false);
-      _pinnedAudio.remove(track.streamUrl);
-      _removeDownload(track.streamUrl);
-    }
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  /// Pins all tracks for an artist for offline playback.
-  Future<void> makeArtistAvailableOffline(
-    Artist artist, {
-    bool requiresWifi = false,
-  }) async {
-    final tracks = await _loadArtistTracksForOffline(artist);
-    for (final track in tracks) {
-      _cancelledOfflineRequests.remove(track.streamUrl);
-      await _cacheStore.setPinnedAudio(track.streamUrl, true);
-      _pinnedAudio.add(track.streamUrl);
-      await _queueDownload(track, requiresWifi: requiresWifi);
-    }
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  /// Removes an artist from offline pinning.
-  Future<void> unpinArtistOffline(Artist artist) async {
-    final tracks = await _loadArtistTracksForOffline(artist);
-    for (final track in tracks) {
-      _cancelledOfflineRequests.add(track.streamUrl);
-      await _cacheStore.setPinnedAudio(track.streamUrl, false);
-      _pinnedAudio.remove(track.streamUrl);
-      _removeDownload(track.streamUrl);
-    }
-    if (_selectedSmartList != null) {
-      unawaited(_loadSmartListTracks(_selectedSmartList!));
-    }
-    unawaited(refreshMediaCacheBytes());
-    notifyListeners();
-  }
-
-  /// Returns whether a track is pinned for offline playback.
-  Future<bool> isTrackPinned(MediaItem track) async {
-    if (_pinnedAudio.isNotEmpty) {
-      return _pinnedAudio.contains(track.streamUrl);
-    }
-    return _cacheStore.isPinnedAudio(track.streamUrl);
-  }
-
-  /// Returns whether any tracks in an album are pinned for offline playback.
-  Future<bool> isAlbumPinned(Album album) async {
-    if (_pinnedAudio.isEmpty) {
-      return false;
-    }
-    final tracks = await _cacheStore.loadAlbumTracks(album.id);
-    if (tracks.isNotEmpty) {
-      return tracks.any((track) => _pinnedAudio.contains(track.streamUrl));
-    }
-    final cachedEntries = await _cacheStore.loadCachedAudioEntries();
-    final albumName = album.name.trim().toLowerCase();
-    return cachedEntries.any(
-      (entry) =>
-          _pinnedAudio.contains(entry.streamUrl) &&
-          entry.album.trim().toLowerCase() == albumName,
-    );
-  }
-
-  /// Returns whether any tracks for an artist are pinned for offline playback.
-  Future<bool> isArtistPinned(Artist artist) async {
-    if (_pinnedAudio.isEmpty) {
-      return false;
-    }
-    final tracks = await _cacheStore.loadArtistTracks(artist.id);
-    if (tracks.isNotEmpty) {
-      return tracks.any((track) => _pinnedAudio.contains(track.streamUrl));
-    }
-    final cachedEntries = await _cacheStore.loadCachedAudioEntries();
-    final artistName = artist.name.trim().toLowerCase();
-    return cachedEntries.any(
-      (entry) =>
-          _pinnedAudio.contains(entry.streamUrl) &&
-          entry.artists.any(
-            (name) => name.trim().toLowerCase() == artistName,
-          ),
-    );
-  }
-
-  /// Returns offline-ready albums based on pinned tracks.
-  Future<List<Album>> loadOfflineAlbums() async {
-    if (_pinnedAudio.isEmpty) {
-      return [];
-    }
-    final cachedEntries = await _cacheStore.loadCachedAudioEntries();
-    final pinnedAlbums = cachedEntries
-        .where((entry) => _pinnedAudio.contains(entry.streamUrl))
-        .map((entry) => entry.album.trim().toLowerCase())
-        .where((name) => name.isNotEmpty)
-        .toSet();
-    if (pinnedAlbums.isEmpty) {
-      return [];
-    }
-    final albums = await _cacheStore.loadAlbums();
-    final offline = albums
-        .where(
-          (album) => pinnedAlbums.contains(album.name.trim().toLowerCase()),
-        )
-        .toList();
-    offline.sort((a, b) => a.name.compareTo(b.name));
-    return offline;
-  }
-
-  /// Returns offline-ready artists based on pinned tracks.
-  Future<List<Artist>> loadOfflineArtists() async {
-    if (_pinnedAudio.isEmpty) {
-      return [];
-    }
-    final cachedEntries = await _cacheStore.loadCachedAudioEntries();
-    final pinnedArtists = <String>{};
-    for (final entry in cachedEntries) {
-      if (!_pinnedAudio.contains(entry.streamUrl)) {
-        continue;
-      }
-      for (final artist in entry.artists) {
-        final normalized = artist.trim().toLowerCase();
-        if (normalized.isNotEmpty) {
-          pinnedArtists.add(normalized);
-        }
-      }
-    }
-    if (pinnedArtists.isEmpty) {
-      return [];
-    }
-    final artists = await _cacheStore.loadArtists();
-    final offline = artists
-        .where((artist) =>
-            pinnedArtists.contains(artist.name.trim().toLowerCase()))
-        .toList();
-    offline.sort((a, b) => a.name.compareTo(b.name));
-    return offline;
-  }
-
-  /// Returns offline-ready playlists based on pinned tracks.
-  Future<List<Playlist>> loadOfflinePlaylists() async {
-    if (_pinnedAudio.isEmpty) {
-      return [];
-    }
-    final playlists = await _cacheStore.loadPlaylists();
-    final offline = <Playlist>[];
-    for (final playlist in playlists) {
-      final tracks = await _cacheStore.loadPlaylistTracks(playlist.id);
-      if (tracks.any((track) => _pinnedAudio.contains(track.streamUrl))) {
-        offline.add(playlist);
-      }
-    }
-    offline.sort((a, b) => a.name.compareTo(b.name));
-    return offline;
-  }
-
-  /// Returns offline-ready tracks based on pinned audio.
-  Future<List<MediaItem>> loadOfflineTracks() async {
-    if (_pinnedAudio.isEmpty) {
-      return [];
-    }
-    final cached = await _cacheStore.loadCachedAudioEntries();
-    final offline = cached
-        .where((entry) => _pinnedAudio.contains(entry.streamUrl))
-        .map(_mediaItemFromCachedEntry)
-        .toList();
-    return offline;
-  }
-
   String _extractStreamItemId(String streamUrl) {
     final uri = Uri.tryParse(streamUrl);
     if (uri == null) {
@@ -4026,91 +1210,6 @@ class AppState extends ChangeNotifier {
     _pinnedCacheBytesNotifier.dispose();
     unawaited(_playback.dispose());
     super.dispose();
-  }
-
-  Future<void> _loadCachedLibrary() async {
-    _playlists = await _cacheStore.loadPlaylists();
-    _featuredTracks = await _cacheStore.loadFeaturedTracks();
-    _albums = await _cacheStore.loadAlbums();
-    _artists = await _cacheStore.loadArtists();
-    _genres = await _cacheStore.loadGenres();
-    _favoriteAlbums = await _cacheStore.loadFavoriteAlbums();
-    _favoriteArtists = await _cacheStore.loadFavoriteArtists();
-    _favoriteTracks = await _cacheStore.loadFavoriteTracks();
-    _recentTracks = await _cacheStore.loadRecentTracks();
-    _playHistory = await _cacheStore.loadPlayHistory();
-    _libraryStats = await _cacheStore.loadLibraryStats();
-    notifyListeners();
-  }
-
-  Future<void> _applyOfflineModeData() async {
-    _isLoadingLibrary = true;
-    clearSearch(notify: false);
-    notifyListeners();
-    _pinnedAudio = await _cacheStore.loadPinnedAudio();
-    final offlineTracks = await loadOfflineTracks();
-    final offlineAlbums = await loadOfflineAlbums();
-    final offlineArtists = await loadOfflineArtists();
-    final offlinePlaylists = await loadOfflinePlaylists();
-    _genres = await _cacheStore.loadGenres();
-    _libraryStats = await _cacheStore.loadLibraryStats();
-    final offlineAlbumIds = offlineAlbums.map((album) => album.id).toSet();
-    final offlineArtistIds = offlineArtists.map((artist) => artist.id).toSet();
-    _libraryTracks = offlineTracks;
-    _tracksOffset = offlineTracks.length;
-    _hasMoreTracks = false;
-    _isLoadingTracks = false;
-    _libraryTracksFromOfflineSnapshot = true;
-    _featuredTracks = offlineTracks;
-    _recentTracks = offlineTracks;
-    _albums = offlineAlbums;
-    _artists = offlineArtists;
-    _playlists = offlinePlaylists;
-    final cachedFavorites = await _cacheStore.loadFavoriteTracks();
-    _favoriteTracks = _filterPinnedTracks(cachedFavorites);
-    final cachedFavoriteAlbums = await _cacheStore.loadFavoriteAlbums();
-    _favoriteAlbums = cachedFavoriteAlbums
-        .where((album) => offlineAlbumIds.contains(album.id))
-        .toList();
-    final cachedFavoriteArtists = await _cacheStore.loadFavoriteArtists();
-    _favoriteArtists = cachedFavoriteArtists
-        .where((artist) => offlineArtistIds.contains(artist.id))
-        .toList();
-    await _refreshSelectedDetailsForOfflineMode();
-    _jumpInTrack = _randomFromList(offlineTracks);
-    _jumpInAlbum = _randomFromList(offlineAlbums);
-    _jumpInArtist = _randomFromList(offlineArtists);
-    _lastJumpInRefreshAt = DateTime.now();
-    if (_selectedSmartList != null) {
-      _smartListTracks = _buildSmartListTracks(_selectedSmartList!);
-    }
-    _isLoadingLibrary = false;
-    notifyListeners();
-  }
-
-  Future<void> _refreshSelectedDetailsForOfflineMode() async {
-    if (_selectedPlaylist != null) {
-      _playlistTracks =
-          await _cacheStore.loadPlaylistTracks(_selectedPlaylist!.id);
-    }
-    if (_selectedAlbum != null) {
-      final cached = await _cacheStore.loadAlbumTracks(_selectedAlbum!.id);
-      final filtered = _filterPinnedTracks(cached);
-      _albumTracks = filtered.isNotEmpty
-          ? filtered
-          : await _offlineTracksForAlbum(_selectedAlbum!);
-    }
-    if (_selectedArtist != null) {
-      final cached = await _cacheStore.loadArtistTracks(_selectedArtist!.id);
-      final filtered = _filterPinnedTracks(cached);
-      _artistTracks = filtered.isNotEmpty
-          ? filtered
-          : await _offlineTracksForArtist(_selectedArtist!);
-    }
-    if (_selectedGenre != null) {
-      final cached = await _cacheStore.loadGenreTracks(_selectedGenre!.id);
-      _genreTracks = _filterPinnedTracks(cached);
-    }
   }
 
   void _bindPlayback() {
@@ -4574,373 +1673,6 @@ class AppState extends ChangeNotifier {
         duration: _duration,
       ),
     );
-  }
-
-  Future<void> _loadAlbums() async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      _isLoadingLibrary = true;
-      notifyListeners();
-      final albums = await _client.fetchAlbums();
-      _albums = albums;
-      await _cacheStore.saveAlbums(albums);
-    } catch (_) {
-      // Use cached results when available.
-    } finally {
-      _isLoadingLibrary = false;
-      notifyListeners();
-    }
-  }
-
-  Future<void> _loadArtists() async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      _isLoadingLibrary = true;
-      notifyListeners();
-      final artists = await _client.fetchArtists();
-      _artists = artists;
-      await _cacheStore.saveArtists(artists);
-    } catch (_) {
-      // Use cached results when available.
-    } finally {
-      _isLoadingLibrary = false;
-      notifyListeners();
-    }
-  }
-
-  Future<void> _loadGenres() async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      _isLoadingLibrary = true;
-      notifyListeners();
-      final genres = await _client.fetchGenres();
-      _genres = genres;
-      await _cacheStore.saveGenres(genres);
-    } catch (_) {
-      // Use cached results when available.
-    } finally {
-      _isLoadingLibrary = false;
-      notifyListeners();
-    }
-  }
-
-  Future<void> _loadFavoriteAlbums() async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      _isLoadingLibrary = true;
-      notifyListeners();
-      final albums = await _client.fetchFavoriteAlbums();
-      _favoriteAlbums = albums;
-      await _cacheStore.saveFavoriteAlbums(albums);
-    } catch (_) {
-      // Use cached results when available.
-    } finally {
-      _isLoadingLibrary = false;
-      notifyListeners();
-      if (_autoDownloadFavoritesEnabled && _autoDownloadFavoriteAlbums) {
-        unawaited(_prefetchFavoriteDownloads(albumsOnly: true));
-      }
-    }
-  }
-
-  Future<void> _loadFavoriteArtists() async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      _isLoadingLibrary = true;
-      notifyListeners();
-      final artists = await _client.fetchFavoriteArtists();
-      if (artists.isNotEmpty || _favoriteArtists.isEmpty) {
-        _favoriteArtists = artists;
-        await _cacheStore.saveFavoriteArtists(artists);
-      }
-    } catch (_) {
-      // Use cached results when available.
-    } finally {
-      _isLoadingLibrary = false;
-      notifyListeners();
-      if (_autoDownloadFavoritesEnabled && _autoDownloadFavoriteArtists) {
-        unawaited(_prefetchFavoriteDownloads(artistsOnly: true));
-      }
-    }
-  }
-
-  Future<void> _loadFavoriteTracks() async {
-    if (_session == null) {
-      return;
-    }
-    if (_offlineMode) {
-      return;
-    }
-    try {
-      _isLoadingLibrary = true;
-      notifyListeners();
-      final tracks = await _client.fetchFavoriteTracks();
-      _favoriteTracks = tracks;
-      await _cacheStore.saveFavoriteTracks(tracks);
-    } catch (_) {
-      // Use cached results when available.
-    } finally {
-      _isLoadingLibrary = false;
-      notifyListeners();
-      if (_autoDownloadFavoritesEnabled && _autoDownloadFavoriteTracks) {
-        unawaited(_prefetchFavoriteDownloads(tracksOnly: true));
-      }
-    }
-  }
-
-  void _applyAlbumFavoriteLocal(Album album, bool isFavorite) {
-    if (isFavorite) {
-      if (!_favoriteAlbums.any((item) => item.id == album.id)) {
-        _favoriteAlbums = [..._favoriteAlbums, album];
-      }
-    } else {
-      _favoriteAlbums =
-          _favoriteAlbums.where((item) => item.id != album.id).toList();
-    }
-    _favoriteAlbums.sort((a, b) => a.name.compareTo(b.name));
-  }
-
-  void _applyArtistFavoriteLocal(Artist artist, bool isFavorite) {
-    if (isFavorite) {
-      if (!_favoriteArtists.any((item) => item.id == artist.id)) {
-        _favoriteArtists = [..._favoriteArtists, artist];
-      }
-    } else {
-      _favoriteArtists =
-          _favoriteArtists.where((item) => item.id != artist.id).toList();
-    }
-    _favoriteArtists.sort((a, b) => a.name.compareTo(b.name));
-  }
-
-  void _applyTrackFavoriteLocal(MediaItem track, bool isFavorite) {
-    if (isFavorite) {
-      if (!_favoriteTracks.any((item) => item.id == track.id)) {
-        _favoriteTracks = [..._favoriteTracks, track];
-      }
-    } else {
-      _favoriteTracks =
-          _favoriteTracks.where((item) => item.id != track.id).toList();
-    }
-    _favoriteTracks.sort((a, b) => a.title.compareTo(b.title));
-  }
-
-  Future<void> _syncAlbumFavoriteOffline(
-    Album album,
-    bool isFavorite,
-  ) async {
-    if (!_autoDownloadFavoritesEnabled || !_autoDownloadFavoriteAlbums) {
-      return;
-    }
-    if (isFavorite) {
-      await makeAlbumAvailableOffline(
-        album,
-        requiresWifi: _autoDownloadFavoritesWifiOnly,
-      );
-      return;
-    }
-    await unpinAlbumOffline(album);
-  }
-
-  Future<void> _syncArtistFavoriteOffline(
-    Artist artist,
-    bool isFavorite,
-  ) async {
-    if (!_autoDownloadFavoritesEnabled || !_autoDownloadFavoriteArtists) {
-      return;
-    }
-    if (isFavorite) {
-      await makeArtistAvailableOffline(
-        artist,
-        requiresWifi: _autoDownloadFavoritesWifiOnly,
-      );
-      return;
-    }
-    await unpinArtistOffline(artist);
-  }
-
-  Future<void> _syncTrackFavoriteOffline(
-    MediaItem track,
-    bool isFavorite,
-  ) async {
-    if (!_autoDownloadFavoritesEnabled || !_autoDownloadFavoriteTracks) {
-      return;
-    }
-    if (isFavorite) {
-      await makeTrackAvailableOffline(
-        track,
-        requiresWifi: _autoDownloadFavoritesWifiOnly,
-      );
-      return;
-    }
-    await unpinTrackOffline(track);
-  }
-
-  /// Updates the favorite status for an album.
-  Future<String?> setAlbumFavorite(Album album, bool isFavorite) async {
-    if (_favoriteAlbumUpdatesInFlight.contains(album.id)) {
-      return null;
-    }
-    final wasFavorite = isFavoriteAlbum(album.id);
-    if (wasFavorite == isFavorite) {
-      return null;
-    }
-    _favoriteAlbumUpdatesInFlight.add(album.id);
-    _applyAlbumFavoriteLocal(album, isFavorite);
-    notifyListeners();
-    if (_offlineMode) {
-      await _cacheStore.saveFavoriteAlbums(_favoriteAlbums);
-      if (_selectedSmartList != null) {
-        unawaited(_loadSmartListTracks(_selectedSmartList!));
-      }
-      _favoriteAlbumUpdatesInFlight.remove(album.id);
-      notifyListeners();
-      return null;
-    }
-    try {
-      await _client.setFavorite(itemId: album.id, isFavorite: isFavorite);
-      await _cacheStore.saveFavoriteAlbums(_favoriteAlbums);
-      if (_selectedSmartList != null) {
-        unawaited(_loadSmartListTracks(_selectedSmartList!));
-      }
-      unawaited(_syncAlbumFavoriteOffline(album, isFavorite));
-      return null;
-    } catch (error) {
-      _applyAlbumFavoriteLocal(album, wasFavorite);
-      await _cacheStore.saveFavoriteAlbums(_favoriteAlbums);
-      return _requestErrorMessage(
-        error,
-        fallback: 'Unable to update album favorite.',
-      );
-    } finally {
-      _favoriteAlbumUpdatesInFlight.remove(album.id);
-      notifyListeners();
-    }
-  }
-
-  /// Updates the favorite status for an artist.
-  Future<String?> setArtistFavorite(Artist artist, bool isFavorite) async {
-    if (_favoriteArtistUpdatesInFlight.contains(artist.id)) {
-      return null;
-    }
-    final wasFavorite = isFavoriteArtist(artist.id);
-    if (wasFavorite == isFavorite) {
-      return null;
-    }
-    _favoriteArtistUpdatesInFlight.add(artist.id);
-    _applyArtistFavoriteLocal(artist, isFavorite);
-    notifyListeners();
-    if (_offlineMode) {
-      await _cacheStore.saveFavoriteArtists(_favoriteArtists);
-      if (_selectedSmartList != null) {
-        unawaited(_loadSmartListTracks(_selectedSmartList!));
-      }
-      _favoriteArtistUpdatesInFlight.remove(artist.id);
-      notifyListeners();
-      return null;
-    }
-    try {
-      await _client.setFavorite(itemId: artist.id, isFavorite: isFavorite);
-      await _cacheStore.saveFavoriteArtists(_favoriteArtists);
-      if (_selectedSmartList != null) {
-        unawaited(_loadSmartListTracks(_selectedSmartList!));
-      }
-      unawaited(_syncArtistFavoriteOffline(artist, isFavorite));
-      final confirmed = await _client.fetchFavoriteState(artist.id);
-      if (confirmed != null && confirmed != isFavorite) {
-        _applyArtistFavoriteLocal(artist, wasFavorite);
-        await _cacheStore.saveFavoriteArtists(_favoriteArtists);
-        notifyListeners();
-        return 'Server did not update artist favorite.';
-      }
-      return null;
-    } catch (error) {
-      _applyArtistFavoriteLocal(artist, wasFavorite);
-      await _cacheStore.saveFavoriteArtists(_favoriteArtists);
-      return _requestErrorMessage(
-        error,
-        fallback: 'Unable to update artist favorite.',
-      );
-    } finally {
-      _favoriteArtistUpdatesInFlight.remove(artist.id);
-      notifyListeners();
-    }
-  }
-
-  /// Updates the favorite status for a track.
-  Future<String?> setTrackFavorite(MediaItem track, bool isFavorite) async {
-    if (_favoriteTrackUpdatesInFlight.contains(track.id)) {
-      return null;
-    }
-    final wasFavorite = isFavoriteTrack(track.id);
-    if (wasFavorite == isFavorite) {
-      return null;
-    }
-    _favoriteTrackUpdatesInFlight.add(track.id);
-    _applyTrackFavoriteLocal(track, isFavorite);
-    notifyListeners();
-    if (_offlineMode) {
-      await _cacheStore.saveFavoriteTracks(_favoriteTracks);
-      if (_selectedSmartList != null) {
-        unawaited(_loadSmartListTracks(_selectedSmartList!));
-      }
-      _favoriteTrackUpdatesInFlight.remove(track.id);
-      notifyListeners();
-      return null;
-    }
-    try {
-      await _client.setFavorite(itemId: track.id, isFavorite: isFavorite);
-      await _cacheStore.saveFavoriteTracks(_favoriteTracks);
-      unawaited(_syncTrackFavoriteOffline(track, isFavorite));
-      return null;
-    } catch (error) {
-      _applyTrackFavoriteLocal(track, wasFavorite);
-      await _cacheStore.saveFavoriteTracks(_favoriteTracks);
-      return _requestErrorMessage(
-        error,
-        fallback: 'Unable to update track favorite.',
-      );
-    } finally {
-      _favoriteTrackUpdatesInFlight.remove(track.id);
-      notifyListeners();
-    }
-  }
-
-  /// Clears album, artist, and genre selections.
-  void clearBrowseSelection({bool notify = true}) {
-    _selectedAlbum = null;
-    _selectedArtist = null;
-    _selectedGenre = null;
-    _albumTracks = [];
-    _artistTracks = [];
-    _genreTracks = [];
-    if (notify) {
-      notifyListeners();
-    }
   }
 
   T? _randomFromList<T>(List<T> items) {
